% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{activations.cifti}
\alias{activations.cifti}
\title{Activations of (s)tICA for CIFTI data}
\usage{
activations.cifti(
  result,
  spatial_model = NULL,
  u = 0,
  alpha = 0.01,
  type = ">",
  method_p = "BH",
  verbose = FALSE,
  which.ICs = NULL,
  deviation = FALSE
)
}
\arguments{
\item{result}{Result of templateICA.cifti model call}

\item{spatial_model}{Should spatial model result be used, if available?  If FALSE, will use standard template ICA result. If NULL, use spatial model result if available.}

\item{u}{Activation threshold, default = 0}

\item{alpha}{Significance level for joint PPM, default = 0.1}

\item{type}{Type of region.  Default is '>' (positive excursion region).}

\item{method_p}{Type of multiple comparisons correction to use for p-values for standard template ICA, or NULL for no correction.  See \code{help(p.adjust)}.}

\item{verbose}{If \code{TRUE}, display progress of algorithm. Default: \code{FALSE}.}

\item{which.ICs}{Indices of ICs for which to identify activations.  If NULL, use all ICs.}

\item{deviation}{If \code{TRUE}. identify significant deviations from the template mean, rather than significant areas of engagement}
}
\value{
A list containing activation maps for each IC and the joint and marginal PPMs for each IC.
}
\description{
Identify areas of activation in each independent component map
}
\keyword{internal}
