% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations_utils.R
\name{add_str}
\alias{add_str}
\title{Combine additive terms in string}
\usage{
add_str(a, b)
}
\arguments{
\item{a, b}{The two terms. Each is a length-one character vector.}
}
\value{
The result as a length-one character vector.
}
\description{
Combine two terms with "+" or "-" in a string
}
\keyword{internal}
