% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_templateICA.R, R/VB_FCtemplateICA.R
\name{bdiag_m}
\alias{bdiag_m}
\title{Block diagonal matrix}
\usage{
bdiag_m(lmat)

bdiag_m(lmat)
}
\arguments{
\item{lmat}{list(\if{html}{\out{<mat1>}}, \if{html}{\out{<mat2>}}, ....., <mat_N>)  where each mat_j is a  k x k 'matrix'}
}
\value{
A sparse matrix obtained by combining the arguments into a block diagonal matrix

a sparse (N\emph{k x N}k) matrix of class  \code{"\linkS4class{dgCMatrix}"}.
}
\description{
Code for function provided in examples of bdiag function from Matrix package
}
\details{
Construct block diagonal matrix of many small blocks
}
\keyword{internal}
