% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{bdiag_m2}
\alias{bdiag_m2}
\title{Fast version of Matrix :: .bdiag() -- for the case of \emph{many identical}  (k x k) matrices:}
\usage{
bdiag_m2(mat, N)
}
\arguments{
\item{mat}{a  k x k 'matrix'}

\item{N}{how many times to repeat \code{mat}}
}
\value{
a sparse (N\emph{k x N}k) matrix of class  \code{"\linkS4class{dgCMatrix}"}.
}
\description{
Fast version of Matrix :: .bdiag() -- for the case of \emph{many identical}  (k x k) matrices:
}
