% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{center_Bcols_Param}
\alias{center_Bcols_Param}
\title{center_Bcols}
\arguments{
\item{center_Bcols}{Center BOLD across columns (each image)? Default:
\code{FALSE} (not recommended).}
}
\description{
center_Bcols
}
\keyword{internal}
