% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_req_ifti_pkg}
\alias{check_req_ifti_pkg}
\title{Check required packages for the data format}
\usage{
check_req_ifti_pkg(FORMAT)
}
\arguments{
\item{FORMAT}{The data FORMAT}
}
\value{
\code{NULL}, invisibly
}
\description{
Check required packages for the data format
}
\keyword{internal}
