% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{Chol_samp_fun}
\alias{Chol_samp_fun}
\title{Cholesky-based FC sampling}
\usage{
Chol_samp_fun(Chol_vals, p, M, chol_diag, chol_offdiag, Chol_mat_blank)
}
\arguments{
\item{Chol_vals}{Matrix of Cholesky factorizations (upper triangular values) for all template sessions (nN*nM x nChol)}

\item{p}{Pivot/reordering applied to FC matrices prior to Cholesky factorization}

\item{M}{Number of samples to draw}

\item{chol_diag}{Indices of diagonal upper triangular elements}

\item{chol_offdiag}{Indices of off-diagonal upper triangular elements}

\item{Chol_mat_blank}{A nLxnL matrix indexing the upper triangular elements}
}
\description{
Cholesky-based FC sampling
}
