% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colCenter}
\alias{colCenter}
\title{Center cols}
\usage{
colCenter(X)
}
\arguments{
\item{X}{The data matrix. Its columns will be centered}
}
\value{
The centered data
}
\description{
Efficiently center columns of a matrix. (Faster than \code{scale})
}
\keyword{internal}
