% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_FCtemplateICA.R
\name{compute_LL_FC}
\alias{compute_LL_FC}
\title{Compute LL for EM FC Template ICA}
\usage{
compute_LL_FC(
  post_sums,
  theta_new,
  prior_params,
  template_FC,
  ntime,
  nICs,
  nvox,
  BOLD,
  verbose = FALSE
)
}
\arguments{
\item{post_sums}{Posterior sums from \code{Gibbs_AS_posteriorCPP}}

\item{theta_new}{List with \code{tau_sq}, \code{alpha}, and \code{G}.}

\item{template_FC}{FC template, for \code{nu} and \code{psi}}

\item{ntime, nICs, nvox}{Number of timepoints, ICs, and voxels}

\item{BOLD}{(\eqn{V \times T} matrix) preprocessed fMRI data}

\item{verbose}{Print updates?}
}
\value{
The expected log posterior at the current values
}
\description{
Compute the expected log posterior for EM FC Template ICA
}
\keyword{internal}
