% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateTheta.R
\name{compute_Sigma_inv}
\alias{compute_Sigma_inv}
\title{Creates precision matrix for S and SPDE matrices (F, G and GFinvG)}
\usage{
compute_Sigma_inv(spde, kappa, template_var, C1 = 1/(4 * pi))
}
\arguments{
\item{spde}{Object of class SPDE representing spatial process}

\item{kappa}{Current estimates of SPDE parameter kappa for each latent field}

\item{template_var}{QxV matrix of template variances}

\item{C1}{Constant, equal to 1/(4*pi) for a 2-dimensional field with alpha=2}
}
\value{
A list containing Sigma inverse and SPDE matrices
}
\description{
Creates precision matrix for S and SPDE matrices (F, G and GFinvG)
}
