% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{detrend_DCT_Param}
\alias{detrend_DCT_Param}
\title{detrend_DCT}
\arguments{
\item{detrend_DCT}{Detrend the data? This is an integer number of DCT bases
to use for detrending. If \code{0} (default), do not detrend.}
}
\description{
detrend_DCT
}
\keyword{internal}
