% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduce.R
\name{dim_reduce}
\alias{dim_reduce}
\title{PCA-based Dimension Reduction}
\usage{
dim_reduce(X, q = NULL)
}
\arguments{
\item{X}{nxp data matrix, centered by columns and rows (columns are actually all that matter, but MATLAB implementation of Minka method also assumes rows have been centered (implicit in use of cov function))}

\item{q}{Number of latent dimensions to estimate}
}
\value{
A list containing the U and D matrices from SVD, prewhitening matrix H and its inverse, the estimated noise variance, and the dimensionality
}
\description{
Estimates the dimension-reducing (aka prewhitening) matrix H based on a certain dimensionality. If dimensionality not specified, estimated using the method described in Minka (2008).
}
