% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_reg.R
\name{dual_reg_fast}
\alias{dual_reg_fast}
\title{Dual Regression Fast}
\usage{
dual_reg_fast(dat, GICA, scale = FALSE)
}
\arguments{
\item{dat}{Subject-level fMRI data (\eqn{VxT})}

\item{GICA}{\code{GICA \%*\% solve(crossprod(GICA))} where \code{GICA} are the 
column-centered group-level independent components (\eqn{VxQ})}

\item{scale}{A logical value indicating whether the fMRI timeseries should 
be scaled by the image standard deviation.}
}
\value{
A list containing the subject-level independent components \strong{S} (\eqn{VxQ}), 
 and subject-level mixing matrix \strong{A} (\eqn{TxQ}).
}
\description{
The GICA term is pre-computed, to avoid repeated computations when this
 function is called in a for loop. Specifically, \code{GICA} here is
 \code{GICA \%*\% solve(crossprod(GICA))} in \code{\link{dual_reg}}.
}
\keyword{internal}
