% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{ELBO}
\alias{ELBO}
\title{ELBO}
\usage{
ELBO(
  mu_S,
  cov_S,
  mu_A,
  cov_A,
  mu_alpha,
  cov_alpha,
  mu_G,
  psi_G,
  nu_G,
  mu_tau2,
  alpha,
  beta,
  template_mean,
  template_var,
  BOLD,
  ntime,
  nICs,
  nvox
)
}
\arguments{
\item{mu_S, cov_S}{(nICs, nvox) and (nICs, nICs, nvox)}

\item{mu_A, cov_A}{TxQ, (QxQ) common across T}

\item{mu_alpha, cov_alpha}{Q, QxQ}

\item{mu_G, psi_G, nu_G}{QxQ, QxQ, 1}

\item{mu_tau2, alpha, beta}{#V, 1, V}

\item{template_mean, template_var}{The templates}

\item{BOLD, ntime, nICs, nvox}{...}
}
\value{
...
}
\description{
Does ELBO ...
}
\keyword{internal}
