% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_algorithm.R
\name{EM_spatialICA}
\alias{EM_spatialICA}
\title{EM Algorithm for Spatial Template ICA}
\usage{
EM_spatialICA(tempICmean, tempICvar, spde, Y, theta0, C_diag,
  max_iter = 100, epsilon = 0.01)
}
\arguments{
\item{tempICmean}{(QxV matrix) mean of each IC in template}

\item{tempICvar}{(QxV matrix) between-subject variance of each IC in template}

\item{spde}{(inla.spde object) SPDE object containing triangular mesh}

\item{Y}{(QxV matrix) dimension-reduced fMRI data}

\item{theta0}{(list) initial guess at parameter values: A (QxQ mixing matrix), nu0_sq (residual variance from first level) and kappa (SPDE smoothness parameter for each IC map)}

\item{C_diag}{(Qx1) diagonal elements of matrix proportional to residual variance.}

\item{max_iter}{maximum number of EM iterations}

\item{epsilon}{smallest proportion change between iterations (e.g. .001)}
}
\value{
A list with 4 elements: theta (list of final parameter estimates), subICmean (estimates of subject-level ICs), subICvar (variance of subject-level ICs), and success (flag indicating convergence (\code{TRUE}) or not (\code{FALSE}))
}
\description{
Implements the expectation-maximization (EM) algorithm described in Mejia et al. (2019+) for estimating the subject-level ICs and unknown parameters in the spatial template ICA model.
}
\details{
If original fMRI timeseries has covariance \eqn{\sigma^2 I_T}, the prewhitened timeseries achieved by premultiplying by (QxT) matrix \eqn{H} from PCA has diagonal covariance \eqn{\sigma^2HH'}, so C_diag is \eqn{diag(HH')}.
}
