% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{estimate.ESS}
\alias{estimate.ESS}
\title{Estimation of effective sample size}
\usage{
estimate.ESS(mesh, Y, ind = NULL, trace = FALSE)
}
\arguments{
\item{mesh}{INLA mesh}

\item{Y}{data}

\item{ind}{index of the data locations in the mesh}
}
\value{
Estimate of the effective sample size
}
\description{
Estimation of effective sample size
}
\details{
The functions computes the effective sample size as 1^T Q.corr 1 where Q.corr is
the inverse of the correlation matrix of a fitted SPDE model with alpha = 1.
If \code{trace=TRUE}, the alternative formula trace(Q^{-1})/trace(Q*Q) is used.
}
