% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.utils.R
\name{estimate_nu}
\alias{estimate_nu}
\title{Universally estimate IW dof parameter nu based on method of moments, so
that no empirical variance is under-estimated}
\usage{
estimate_nu(var_FC, mean_FC)
}
\arguments{
\item{var_FC}{Empirical between-subject variance of covariance matrices (QxQ)}

\item{mean_FC}{Empirical mean of covariance matrices (QxQ)}
}
\value{
estimate for nu
}
\description{
Universally estimate IW dof parameter nu based on method of moments, so
that no empirical variance is under-estimated
}
