% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template.cifti}
\alias{estimate_template.cifti}
\title{Estimate CIFTI template}
\usage{
estimate_template.cifti(
  cifti_fnames,
  cifti_fnames2 = NULL,
  GICA_fname,
  inds = NULL,
  scale = TRUE,
  normA = FALSE,
  brainstructures = c("left", "right"),
  verbose = TRUE,
  out_fname = NULL,
  FC = FALSE
)
}
\arguments{
\item{cifti_fnames}{Vector of file paths of CIFTI-format fMRI timeseries
(*.dtseries.nii) for template estimation}

\item{cifti_fnames2}{Vector of file paths of "retest" CIFTI-format fMRI
timeseries (*.dtseries.nii) for template estimation.  Must be from the same
subjects and in the same order as cifti_fnames.  If none specified, will
create pseudo test-retest data from single session.}

\item{GICA_fname}{File path of CIFTI-format group ICA maps (ending in .d*.nii)}

\item{inds}{Indicators of which group ICs to include in template. If NULL,
use all group ICs.}

\item{scale}{Logical indicating whether BOLD data should be scaled by the
spatial standard deviation before template estimation.}

\item{normA}{Normalize the A matrix (spatial maps)?}

\item{brainstructures}{Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("left","right")} (cortical surface only).}

\item{verbose}{If \code{TRUE}. display progress updates}

\item{out_fname}{(Required if templates are to be resampled to a lower spatial
resolution, usually necessary for spatial template ICA.) The path and base name
prefix of the CIFTI files to write. Will be appended with "_mean.dscalar.nii" for
template mean maps and "_var.dscalar.nii" for template variance maps.}

\item{FC}{If TRUE, include template for functional connectivity}
}
\value{
List of two elements: template mean of class xifti and template
 variance of class xifti
}
\description{
Estimate template for Template or Diagnostic ICA based on CIFTI-format data
}
