% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template.gifti}
\alias{estimate_template.gifti}
\title{Estimate CIFTI template}
\usage{
estimate_template.gifti(
  gifti_fnames,
  GICA_fname,
  inds = NULL,
  scale = TRUE,
  verbose = TRUE,
  out_fname = NULL
)
}
\arguments{
\item{gifti_fnames}{Vector of file paths of GIFTI-format fMRI timeseries
(*.dtseries.nii) for template estimation}

\item{GICA_fname}{File path of CIFTI-format group ICA maps (ending in .d*.nii)}

\item{inds}{Indicators of which group ICs to include in template. If NULL,
use all group ICs.}

\item{scale}{Logical indicating whether BOLD data should be scaled by the
spatial standard deviation before template estimation.}

\item{verbose}{If \code{TRUE}. display progress updates}

\item{out_fname}{(Required if templates are to be resampled to a lower spatial
resolution, usually necessary for spatial template ICA.) The path and base name
prefix of the CIFTI files to write. Will be appended with "_mean.dscalar.nii" for
template mean maps and "_var.dscalar.nii" for template variance maps.}
}
\value{
List of two elements: template mean of class xifti and template
 variance of class xifti
}
\description{
Estimate template for Template or Diagnostic ICA based on CIFTI-format data
}
