% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template_FC}
\alias{estimate_template_FC}
\title{Estimate FC template}
\usage{
estimate_template_FC(FC0)
}
\arguments{
\item{FC0}{The FC estimates from \code{\link{estimate_template}}.}
}
\description{
Estimate FC template
}
\keyword{internal}
