% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template_FC_chol}
\alias{estimate_template_FC_chol}
\title{Estimate FC template from Cholesky decomposition}
\usage{
estimate_template_FC_chol(FC0_chol)
}
\arguments{
\item{FC0_chol}{The FC estimate Cholesky factorizations (upper triangular values) from \code{\link{estimate_template}}.}
}
\description{
Estimate FC template from Cholesky decomposition
}
