% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_template.R
\name{export_template}
\alias{export_template}
\title{Export template}
\usage{
export_template(
  x,
  out_fname = NULL,
  var_method = c("non-negative", "unbiased")
)
}
\arguments{
\item{x}{The result of \code{estimate_template}}

\item{out_fname}{Use \code{NULL} (default) to just return the template
objects directly. Otherwise, use a length-3 character vector of file
path(s) to save the output to:
the mean template, the variance template, and the variance decomposition,
in that order. If one file name is provided, it will be appended with
\code{"_mean.[file_ext]"} for the template mean map,
\code{"_var.[file_ext]"} for the template variance map, and
\code{"_varDecomp.rds"} for the variance decomposition, where \code{[file_ext]}
will be \code{"dscalar.nii"} for CIFTI input, \code{"nii"} for NIFTI input,
and \code{"rds"} for data input.}

\item{var_method}{\code{"non-negative"} (default) or \code{"unbiased"}}
}
\description{
Export the templates (mean and variance) as separate files for
visualization or processing outside of \code{templateICAr}.
}
