% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations_utils.R
\name{format_activation_name}
\alias{format_activation_name}
\title{Format activation name}
\usage{
format_activation_name(u, z, type, deviation, collapse = FALSE)
}
\arguments{
\item{u, z, type, deviation}{See \code{\link{activations}}.}

\item{collapse}{If multiple \code{u} or \code{z} were provided, should just
one name be returned? Default: \code{FALSE}}
}
\description{
Format the name of an activation, given parameters of the statistical test.
}
\keyword{internal}
