% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{GSR_Param}
\alias{GSR_Param}
\title{GSR}
\arguments{
\item{GSR}{Center BOLD across columns (each image)? This
is equivalent to performing global signal regression. Default:
\code{FALSE}.}
}
\description{
GSR
}
\keyword{internal}
