% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{halflogdetX}
\alias{halflogdetX}
\title{Half log determinant}
\usage{
halflogdetX(X)
}
\arguments{
\item{X}{A numeric matrix}
}
\value{
The half log determinant of \code{X}.
}
\description{
Computes half log determinant of \code{X}, by \code{sum(log(diag(chol(X))))}.
}
\keyword{internal}
