% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{hpf_param}
\alias{hpf_param}
\title{hpf}
\arguments{
\item{hpf}{The frequency at which to apply a highpass filter to the data
during pre-processing, in Hertz. Default: \code{0.01} Hertz. Set to \code{0}
to disable the highpass filter.

The highpass filter serves to detrend the data, since low-frequency
variance is associated with noise. Highpass filtering is accomplished by
nuisance regression of discrete cosine transform (DCT) bases.

Note the \code{TR} argument is required for highpass filtering. If
\code{TR} is not provided, \code{hpf} will be ignored.}
}
\description{
hpf
}
\keyword{internal}
