% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_BOLD_format}
\alias{infer_BOLD_format}
\title{Infer fMRI data format}
\usage{
infer_BOLD_format(BOLD, verbose = FALSE)
}
\arguments{
\item{BOLD}{The fMRI data}

\item{verbose}{Print the format? Default: \code{FALSE}.}
}
\value{
The format: \code{"CIFTI"} file path, \code{"xifti"} object,
\code{"NIFTI"} file path, \code{"nifti"} object, or \code{"data"}.
}
\description{
Infer fMRI data format
}
\keyword{internal}
