% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.utils.R
\name{IW_var}
\alias{IW_var}
\title{Compute theoretical Inverse-Wishart variance of covariance matrix elements}
\usage{
IW_var(nu, p, xbar_ij, xbar_ii, xbar_jj)
}
\arguments{
\item{nu}{Inverse Wishart degrees of freedom parameter}

\item{p}{Matrix dimension for IW distribution}

\item{xbar_ij}{Empirical mean of covariance matrices at element (i,j)}

\item{xbar_ii}{Empirical mean of covariance matrices at the ith diagonal element}

\item{xbar_jj}{Empirical mean of covariance matrices at the jth diagonal element}
}
\value{
Theoretical IW variance for covariance element (i,j)
}
\description{
Compute theoretical Inverse-Wishart variance of covariance matrix elements
}
