% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_diagnosticICA.R
\name{LL2_kappa_diagnosticICA}
\alias{LL2_kappa_diagnosticICA}
\title{Compute part of log-likelihood involving kappa (or kappa_q) for numerical optimization}
\usage{
LL2_kappa_diagnosticICA(
  kappa,
  Amat,
  Fmat,
  Gmat,
  GFinvG,
  OplusW,
  u,
  v,
  C1 = 1/(4 * pi),
  Q,
  pr_zy
)
}
\arguments{
\item{kappa}{Value of kappa for which to compute log-likelihood}

\item{Amat}{Mesh projection matrix}

\item{Fmat}{Matrix used in computation of SPDE precision}

\item{Gmat}{Matrix used in computation of SPDE precision}

\item{GFinvG}{Matrix used in computation of SPDE precision}

\item{OplusW}{LIST OF sparse matrices containing estimated values of RHS of trace in part 2 of log-likelihood.  (one list element for each group g)}

\item{u}{LIST OF vectors needed for part 3 of log-likelihood (one list element for each group g)}

\item{v}{LIST OF vectors needed for part 3 of log-likelihood (one list element for each group g)}

\item{C1}{For the unit variance case, \eqn{\tau^2 = C1/\kappa^2}, where \eqn{C1 = 1/(4\pi)} when \eqn{\alpha=2}, \eqn{\nu=1}, \eqn{d=2}}

\item{Q}{Equal to the number of ICs}

\item{pr_zy}{Current posterior probabilities of group membership z}
}
\value{
Value of log-likelihood at logkappa
}
\description{
Compute part of log-likelihood involving kappa (or kappa_q) for numerical optimization
}
\details{
This is the function to be maximized in order to determine the MLE for \eqn{\kappa} or the \eqn{\kappa_q}'s in the M-step of the EM algorithm in spatial
template ICA.  In the model where \eqn{\kappa_q} can be different for each IC \eqn{q}, the optimization function factorizes over the \eqn{\kappa_q}'s.  This function computes
the value of the part of the optimization function pertaining to one of the \eqn{\kappa_q}'s.
}
\keyword{internal}
