% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{log_diwish}
\alias{log_diwish}
\title{Compute the log of the IW density}
\usage{
log_diwish(W, nu, S)
}
\arguments{
\item{W, nu, S}{The arguments ...}
}
\value{
The log of the IW density
}
\description{
Compute the log of the IW density
}
\keyword{internal}
