% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logical_arg}
\alias{logical_arg}
\title{Logical argument}
\usage{
logical_arg(x, name = "x")
}
\arguments{
\item{x}{The argument input}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Make \code{TRUE} or \code{FALSE}
}
\keyword{internal}
