% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_templateICA.R
\name{make_Pmat}
\alias{make_Pmat}
\title{Make permutation matrix}
\usage{
make_Pmat(Q, nvox)
}
\arguments{
\item{Q}{The number of template ICs}

\item{nvox}{The number of spatial locations (\eqn{V})}
}
\value{
P Permutation matrix size \eqn{QVxQV}
}
\description{
Create permutation matrix P to regroup elements of s by locations instead of by ICs
}
\details{
If s=(s1,...,sQ) is grouped by ICs 1,...Q, then Ps=(s(1),...,s(nvox)) is grouped by locations 1,...,nvox
}
\keyword{internal}
