% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maxQ_check}
\alias{maxQ_check}
\title{Check maxQ}
\usage{
maxQ_check(maxQ, L, T)
}
\arguments{
\item{maxQ, L, T}{The args}
}
\value{
\code{maxQ}, clamped to acceptable range of values.
}
\description{
Check maxQ
}
\keyword{internal}
