% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMRItools_devTemp.R
\name{mean_squares}
\alias{mean_squares}
\title{Compute mean squares from variance decomposition}
\usage{
mean_squares(vd)
}
\arguments{
\item{vd}{The variance decomposition}
}
\value{
The mean squares
}
\description{
Compute mean squares from variance decomposition
}
\keyword{internal}
