% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGP.R
\name{MIGP}
\alias{MIGP}
\title{MIGP}
\usage{
MIGP(
  dat,
  datProcFUN,
  checkColCentered = TRUE,
  nM = NULL,
  nP = NULL,
  initW = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{The files or data to compute PCA for, in list format, with each
subject as a separate list entry.}

\item{datProcFUN}{Processes each entry of \code{dat} into a \eqn{T \times V}
matrix (timepoints by locations) which has been column-centered (each
location's timecourse is mean zero) and variance normalized.}

\item{checkColCentered}{Check that each subject's data is column centered
after processing it with \code{datProcFUN}? Default: \code{TRUE}.}

\item{nM}{The size of the "internal" PCA space. Must be larger than the
typical number of timepoints in each individual dataset, \eqn{T}. If
\code{NULL} (default), set to \code{T*2}.}

\item{nP}{The number of final group PCA components to obtain. If
\code{NULL} (default), will be set to \code{nM}.}

\item{initW}{The initial \eqn{D \times V^t} matrix, for example from a previous
call to MIGP. This argument could be used to iteratively build an MIGP result
across many fMRI scans. Default: \code{NULL} (no prior data).}

\item{verbose}{Occasional updates? Default: \code{TRUE}.}

\item{...}{Additional arguments to \code{datProcFUN}}
}
\value{
The dimension-reduced data
}
\description{
Melodic's Incremental Group PCA
}
\details{
https://doi.org/10.1016/j.neuroimage.2014.07.051

Appendix A
}
