% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minka.R
\name{minka}
\alias{minka}
\title{Automated PCA-based Dimension Reduction based on Minka (2008)}
\usage{
minka(data, alpha = 0, beta = 0, eps = 1e-10)
}
\arguments{
\item{data}{Nxd data matrix for N observations of d variables.}

\item{alpha}{Parameter controlling sharpness of prior on eigenvalues and noise variance. Values close to zero lead to uninformative priors.}

\item{beta}{Parameter controlling sharpness of prior on eigenvalues and noise variance. Values close to zero lead to uninformative priors.}

\item{eps}{Value below which sample eigenvalues will be considered equal to zero.}
}
\value{
A list containing the estimated dimensionality, noise variance, and the vector of log probabilities for each dimensionality.
}
\description{
Automated PCA-based Dimension Reduction based on Minka (2008)
}
