% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{normA_Param}
\alias{normA_Param}
\title{normA}
\arguments{
\item{normA}{Scale each IC timeseries (column of \eqn{A}) in the dual
regression estimates? Default: \code{FALSE} (not recommended). Note that the
product \eqn{A \times S} remains the same with either option.}
}
\description{
normA
}
\keyword{internal}
