% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisIC_nreg.R
\name{nuisIC_nreg}
\alias{nuisIC_nreg}
\title{Remove nuisance ICs from data}
\usage{
nuisIC_nreg(
  BOLD,
  DR = NULL,
  template_mean = NULL,
  Q2 = NULL,
  Q2_max = NULL,
  verbose = FALSE
)
}
\arguments{
\item{BOLD}{the \eqn{V} by \eqn{T} data}

\item{DR, template_mean}{We need an initial estimate of the group ICs, so that we can avoid removing them
during removal of the estimated noise ICs. Provide either \code{DR} if dual regression has already
been calculated, or \code{template_mean} if it hasn't. One is required.}

\item{Q2}{the number of nuisance ICs. If \code{NULL} (default) will estimate using PESEL.}

\item{Q2_max}{If \code{Q2} is \code{NULL}, PESEL's estimate will be less than or equal to \code{Q2_max}.
If \code{Q2_max} is \code{NULL}, do not limit PESEL's estimate.}

\item{verbose}{If \code{TRUE}, display progress updates}
}
\value{
The \eqn{V} by \eqn{T} data with the estimated nuisance ICs subtracted from it
}
\description{
Subtract estimated nuisance ICs from data matrix. If the number of nuisance ICs is not provided, 
 will estimate using PESEL (the nuisance ICs are a rotation of the nuisance PCs).
}
\keyword{internal}
