% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduce_PCA.R
\name{PCA}
\alias{PCA}
\title{PCA}
\usage{
PCA(X, Xt = FALSE, Q = NULL, Q_max = 100)
}
\arguments{
\item{X}{\eqn{VxT} fMRI timeseries data matrix, centered by columns and rows 
(columns are actually all that matter, but MATLAB implementation of Minka 
method also assumes rows have been centered (implicit in use of cov function))}

\item{Q}{Number of latent dimensions to estimate. If not specified, 
estimated using PESEL (Sobczyka et al. 2020).}

\item{Q_max}{Maximal number of principal components for automatic 
dimensionality selection with PESEL. Default: \code{100}}
}
\value{
The SVD decomposition
}
\description{
PCA
}
