% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.methods.R
\name{plot.template.cifti}
\alias{plot.template.cifti}
\title{Plot template}
\usage{
\method{plot}{template.cifti}(x, stat = c("both", "mean", "var"), ...)
}
\arguments{
\item{x}{The template from \code{estimate_template.cifti}}

\item{stat}{\code{"mean"}, \code{"var"}, or \code{"both"} (default)}

\item{...}{Additional arguments to \code{view_xifti}}
}
\value{
The plot
}
\description{
Plot template
}
