% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.methods.R
\name{plot.template.matrix}
\alias{plot.template.matrix}
\title{Plot template}
\usage{
\method{plot}{template.matrix}(x, ...)
}
\arguments{
\item{x}{The template from \code{estimate_template.matrix}}

\item{...}{Additional arguments}
}
\value{
The plot
}
\description{
Plot template
}
