% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.methods.R
\name{plot.template.nifti}
\alias{plot.template.nifti}
\title{Plot template}
\usage{
\method{plot}{template.nifti}(x, ...)
}
\arguments{
\item{x}{The template from \code{estimate_template.nifti}}

\item{...}{Additional arguments}
}
\value{
The plot
}
\description{
Plot template
}
