% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.methods.R
\name{plot.template_cifti}
\alias{plot.template_cifti}
\title{Plot template}
\usage{
\method{plot}{template_cifti}(x, stat = c("mean", "var"), ...)
}
\arguments{
\item{x}{The template from \code{estimate_template.cifti}}

\item{stat}{\code{"mean"} (default) or \code{"var"}}

\item{...}{Additional arguments to \code{view_xifti}}
}
\value{
The plot
}
\description{
Plot template
}
