% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.methods.R
\name{plot.templateICA.cifti}
\alias{plot.templateICA.cifti}
\title{Plot template}
\usage{
\method{plot}{templateICA.cifti}(x, stat = c("mean", "se", "both"), ...)
}
\arguments{
\item{x}{The result of \code{templateICA} with CIFTI data}

\item{stat}{\code{"mean"} (default), \code{"se"}, or \code{"both"}}

\item{...}{Additional arguments to \code{view_xifti}}
}
\value{
The plot
}
\description{
Plot template
}
