% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.methods.R
\name{plot.tICA.matrix}
\alias{plot.tICA.matrix}
\title{Plot template}
\usage{
\method{plot}{tICA.matrix}(x, ...)
}
\arguments{
\item{x}{The result of \code{templateICA} with NIFTI data}

\item{...}{Additional arguments}
}
\value{
Nothing, because an error is raised.
}
\description{
This feature is not supported yet.
}
