% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.methods.R
\name{plot.tICA_act.cifti}
\alias{plot.tICA_act.cifti}
\title{Plot activations}
\usage{
\method{plot}{tICA_act.cifti}(x, stat = c("active", "pvals", "pvals_adj", "tstats", "se"), ...)
}
\arguments{
\item{x}{The activations from \code{activations.cifti}}

\item{stat}{\code{"active"} (default), \code{"pvals"}, \code{"pvals_adj"},
\code{"tstats"}, or \code{"vars"}.}

\item{...}{Additional arguments to \code{view_xifti}}
}
\value{
The activations plot
}
\description{
Plot activations
}
