% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.R
\name{plot.tICA_activations}
\alias{plot.tICA_activations}
\title{Plot activations}
\usage{
\method{plot}{tICA_activations}(x, stat = c("active", "pvals", "pvals_adj", "tstats", "vars"), ...)
}
\arguments{
\item{x}{The activations from \code{activations.cifti}}

\item{stat}{\code{"active"} (default), \code{"pvals"}, \code{"pvals_adj"},
\code{"tstats"}, or \code{"vars"}.}

\item{...}{Additional arguments to \code{view_xifti}}
}
\value{
The plot
}
\description{
Plot activations
}
