% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.methods.R
\name{plot_FC}
\alias{plot_FC}
\title{Plot FC template}
\usage{
plot_FC(
  x,
  zlim = c(-1, 1),
  title = NULL,
  cols = c("darkblue", "turquoise", "white", "pink", "red"),
  break_by = 0.5,
  cor = TRUE
)
}
\arguments{
\item{x}{The FC template}

\item{zlim}{Color limits}

\item{title}{Plot title}

\item{cols}{Colors}

\item{break_by}{Color breaks}

\item{cors}{\code{TRUE}}
}
\description{
Plot FC template
}
