% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pw_estimate}
\alias{pw_estimate}
\title{Estimate residual autocorrelation for prewhitening}
\usage{
pw_estimate(A, ar_order, aic = FALSE)
}
\arguments{
\item{A}{Estimated A matrix (T x Q)}

\item{ar_order, aic}{Order of the AR model used to prewhiten the data at each location.
If \code{!aic} (default), the order will be exactly \code{ar_order}. If \code{aic},
the order will be between zero and \code{ar_order}, as determined by the AIC.}
}
\value{
Estimated AR coefficients and residual variance at every vertex
}
\description{
Estimate residual autocorrelation for prewhitening
}
