% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Q2_kappa}
\alias{Q2_kappa}
\title{Computes part of log-likelihood involving kappa (or kappa_q) for numerical optimization}
\usage{
Q2_kappa(logkappa, Fmat, Gmat, GFinvG = NULL, bigTrace1, bigTrace2,
  C1 = 1/(4 * pi), Q = NULL)
}
\arguments{
\item{logkappa}{Value of log(kappa) for which to compute log-likelihood}

\item{Fmat}{A diagonal matrix appearing in SPDE precision}

\item{Gmat}{A sparse neighborhood matrix appearing in SPDE precision}

\item{GFinvG}{(Optional) Pre-computed value of \eqn{GFinvG = Gmat * (1/Fmat)* Gmat} (available from SPDE object)}

\item{bigTrace1}{Value of first large trace term appearing in optimization function (see Details)}

\item{bigTrace2}{Value of second large trace term appearing in optimization function (see Details)}

\item{C1}{For the unit variance case, \eqn{\tau^2 = C1/\kappa^2}, where \eqn{C1 = 1/(4\pi)} when \eqn{\alpha=2}, \eqn{\nu=1}, \eqn{d=2}}

\item{Q}{Equal to the number of ICs for the common smoothness model, or NULL for the IC-specific smoothness model}
}
\value{
Value of log-likelihood at logkappa
}
\description{
Computes part of log-likelihood involving kappa (or kappa_q) for numerical optimization
}
\details{
This is the function to be maximized in order to determine the MLE for \eqn{\kappa} or the \eqn{\kappa_q}'s in the M-step of the EM algorithm in spatial
template ICA.  In the model where \eqn{\kappa_q} can be different for each IC \eqn{q}, the optimization function factorizes over the \eqn{\kappa_q}'s.  This function computes
the value of the part of the optimization function pertaining to one of the \eqn{\kappa_q}'s. The optimization function includes two Trace terms involving
very large matrices, which are passed to this function as \code{bigTrace1} and \code{bigTrace2}.  \code{bigTrace1} is equal to

\deqn{K1q = Trace(Dq_inv * F * Dq_inv * E[delta_q*t(delta_q)])}

and \code{bigTrace2} is equal to

\deqn{K2q = Trace(Dq_inv * G * F_inv * G * Dq_inv * E[delta_q*t(delta_q)]).}

For the common smoothness model, a single kappa value is assumed over all of the IC's. The two Trace terms are then equal to \eqn{\sum_q(K1q)} and \eqn{\sum_q(K2q)}.
}
