% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Q2_kappa_q}
\alias{Q2_kappa_q}
\title{Computes part of log-likelihood involving kappa_q for numerical optimization}
\usage{
Q2_kappa_q(logkappa_q, Fmat, Gmat, GFinvG = NULL, bigTrace1, bigTrace2,
  C1 = 1/(4 * pi))
}
\arguments{
\item{logkappa_q}{Value of log(kappa) for which to compute log-likelihood}

\item{Fmat}{A diagonal matrix appearing in SPDE precision}

\item{Gmat}{A sparse neighborhood matrix appearing in SPDE precision}

\item{GFinvG}{\code{GFinvG = Gmat \%*\% (1/Fmat) \%*\% Gmat} (available directly through SPDE object)}

\item{bigTrace1}{Equal to \code{Trace(Dq_inv * F * Dq_inv * E[delta_q*t(delta_q)])}}

\item{bigTrace2}{Equal to \code{Trace(Dq_inv * G * F_inv * G * Dq_inv * E[delta_q*t(delta_q)])}}

\item{C1}{For the unit variance case, $tau^2 = C1/kappa^2$.  $C1 = 1/(4*pi)$ when $alpha=2$, $nu=1$, $d=2$}
}
\value{
Value of log-likelihood at logkappa_q
}
\description{
Computes part of log-likelihood involving kappa_q for numerical optimization
}
