% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Q2_max_check}
\alias{Q2_max_check}
\title{Check Q2_max}
\usage{
Q2_max_check(Q2_max, nQ, nT)
}
\arguments{
\item{Q2_max, nQ, nT}{The args}
}
\value{
\code{Q2_max}, clamped to acceptable range of values.
}
\description{
Check Q2_max
}
\keyword{internal}
