% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removebs_template.R
\name{removebs_template}
\alias{removebs_template}
\title{Remove brain structure from CIFTI template}
\usage{
removebs_template(x, remove = NULL)
}
\arguments{
\item{x}{The \code{"template.cifti"} object.}

\item{remove}{\code{"cortex_left"}, \code{"cortex_right"}, and/or \code{"subcortical"}.}
}
\description{
Remove a brain structure from a CIFTI template
}
\keyword{internal}
