% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_template.R
\name{resample_template}
\alias{resample_template}
\title{Resample CIFTI template}
\usage{
resample_template(x, resamp_res, verbose = FALSE)
}
\arguments{
\item{x}{The \code{"template.cifti"} object.}

\item{resamp_res}{The new resampling resolution.}

\item{verbose}{Give occasional updates? Default: \code{FALSE}.}
}
\value{
The resampled \code{"template.cifti"} object.
}
\description{
Resample a CIFTI template to a new spatial resolution.
}
