% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_BOLD.R
\name{scale_BOLD}
\alias{scale_BOLD}
\title{Scale BOLD (legacy version of \code{norm_BOLD}) that centers both ways.}
\usage{
scale_BOLD(BOLD, scale = FALSE)
}
\arguments{
\item{BOLD}{fMRI numeric data matrix (\eqn{V \times T})}

\item{scale}{A logical value indicating whether the fMRI timeseries should be scaled by the image standard deviation. Default: \code{FALSE}.}
}
\description{
Scale BOLD (legacy version of \code{norm_BOLD}) that centers both ways.
}
