% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rox_args_docs.R
\name{scale_sm_FWHM_Param}
\alias{scale_sm_FWHM_Param}
\title{scale_sm_FWHM}
\arguments{
\item{scale_sm_FWHM}{Only applies if \code{scale=="local"} and if \code{BOLD}
represents surface data. To smooth the standard deviation estimates used for
local scaling, provide the smoothing FWHM (default: \code{2}). As the FWHM
becomes large, local scaling approaches global scaling.}
}
\description{
scale_sm_FWHM
}
\keyword{internal}
