% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sign_flip}
\alias{sign_flip}
\title{Sign match ICA results}
\usage{
sign_flip(x)
}
\arguments{
\item{x}{The ICA results with entries \code{S} and \code{M}}
}
\value{
\code{x} but with positive skew source signals
}
\description{
Flips all source signal estiamtes (S) to positive skew
}
\keyword{internal}
