% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sign_match}
\alias{sign_match}
\title{Sign matching}
\usage{
sign_match(x)
}
\arguments{
\item{x}{the vector}
}
\value{
\code{x} if the skew of \code{x} is negative, and \code{-x} if
 the skew of \code{x} is positive
}
\description{
Standardize sign of vector by flipping it if the skew is negative.
}
\keyword{internal}
