% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.methods.R
\name{summary.template.data}
\alias{summary.template.data}
\alias{print.summary.template.data}
\alias{print.template.data}
\title{Summarize a \code{"template.data"} object}
\usage{
\method{summary}{template.data}(object, ...)

\method{print}{summary.template.data}(x, ...)

\method{print}{template.data}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"template.data"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The template from \code{estimate_template.cifti}}
}
\description{
Summary method for class \code{"template.data"}
}
