% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.methods.R
\name{summary.template_cifti}
\alias{summary.template_cifti}
\alias{print.summary.template_cifti}
\alias{print.template_cifti}
\title{Summarize a \code{"template_cifti"} object}
\usage{
\method{summary}{template_cifti}(object, ...)

\method{print}{summary.template_cifti}(x, ...)

\method{print}{template_cifti}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"template_cifti"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The template from \code{estimate_template.cifti}}
}
\description{
Summary method for class \code{"template_cifti"}
}
