% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.methods.R
\name{summary.template_data}
\alias{summary.template_data}
\alias{print.summary.template_data}
\alias{print.template_data}
\title{Summarize a \code{"template_data"} object}
\usage{
\method{summary}{template_data}(object, ...)

\method{print}{summary.template_data}(x, ...)

\method{print}{template_data}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"template_data"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The template from \code{estimate_template.cifti}}
}
\description{
Summary method for class \code{"template_data"}
}
