% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.methods.R
\name{summary.templateICA.cifti}
\alias{summary.templateICA.cifti}
\alias{print.summary.templateICA.cifti}
\alias{print.templateICA.cifti}
\title{Summarize a \code{"templateICA.cifti"} object}
\usage{
\method{summary}{templateICA.cifti}(object, ...)

\method{print}{summary.templateICA.cifti}(x, ...)

\method{print}{templateICA.cifti}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"templateICA.cifti"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The result of \code{templateICA} with CIFTI data}
}
\description{
Summary method for class \code{"templateICA.cifti"}
}
