% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.methods.R
\name{summary.tICA.cifti}
\alias{summary.tICA.cifti}
\alias{print.summary.tICA.cifti}
\alias{print.tICA.cifti}
\title{Summarize a \code{"tICA.cifti"} object}
\usage{
\method{summary}{tICA.cifti}(object, ...)

\method{print}{summary.tICA.cifti}(x, ...)

\method{print}{tICA.cifti}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"tICA.cifti"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The result of \code{templateICA} with CIFTI data}
}
\value{
A list summarizing of the results of the templateICA analysis.

Nothing, invisibly.

Nothing, invisibly.
}
\description{
Summary method for class \code{"tICA.cifti"}
}
