% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.methods.R
\name{summary.tICA.matrix}
\alias{summary.tICA.matrix}
\alias{print.summary.tICA.matrix}
\alias{print.tICA.matrix}
\title{Summarize a \code{"tICA.matrix"} object}
\usage{
\method{summary}{tICA.matrix}(object, ...)

\method{print}{summary.tICA.matrix}(x, ...)

\method{print}{tICA.matrix}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"tICA.matrix"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The template from \code{estimate_template.cifti}}
}
\value{
A list summarizing of the results of the templateICA analysis.

Nothing, invisibly.

Nothing, invisibly.
}
\description{
Summary method for class \code{"tICA.matrix"}
}
