% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.methods.R
\name{summary.tICA}
\alias{summary.tICA}
\alias{print.summary.tICA}
\alias{print.tICA}
\title{Summarize a \code{"tICA"} object}
\usage{
\method{summary}{tICA}(object, ...)

\method{print}{summary.tICA}(x, ...)

\method{print}{tICA}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"tICA"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The template from \code{estimate_template.cifti}}
}
\description{
Summary method for class \code{"tICA"}
}
