% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.methods.R
\name{summary.tICA_act.cifti}
\alias{summary.tICA_act.cifti}
\alias{print.summary.tICA_act.cifti}
\alias{print.tICA_act.cifti}
\title{Summarize a \code{"tICA_act.cifti"} object}
\usage{
\method{summary}{tICA_act.cifti}(object, ...)

\method{print}{summary.tICA_act.cifti}(x, ...)

\method{print}{tICA_act.cifti}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"tICA_act.cifti"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The activations from \code{activations.cifti}}
}
\value{
A list summarizing the data and results for the activations analysis.

Nothing, invisibly.

Nothing, invisibly.
}
\description{
Summary method for class \code{"tICA_act.cifti"}
}
