% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.methods.R
\name{summary.tICA_act.matrix}
\alias{summary.tICA_act.matrix}
\alias{print.summary.tICA_act.matrix}
\alias{print.tICA_act.matrix}
\title{Summarize a \code{"tICA_act.matrix"} object}
\usage{
\method{summary}{tICA_act.matrix}(object, ...)

\method{print}{summary.tICA_act.matrix}(x, ...)

\method{print}{tICA_act.matrix}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"tICA_act.matrix"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The activations from \code{activations}}
}
\value{
A list summarizing the data and results for the activations analysis.

Nothing, invisibly.

Nothing, invisibly.
}
\description{
Summary method for class \code{"tICA_act.matrix"}
}
