% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activations.methods.R
\name{summary.tICA_act}
\alias{summary.tICA_act}
\alias{print.summary.tICA_act}
\alias{print.tICA_act}
\title{Summarize a \code{"tICA_act"} object}
\usage{
\method{summary}{tICA_act}(object, ...)

\method{print}{summary.tICA_act}(x, ...)

\method{print}{tICA_act}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"tICA_act"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The activations from \code{activations}}
}
\description{
Summary method for class \code{"tICA_act"}
}
