% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.cifti.R
\name{summary.tICA_activations}
\alias{summary.tICA_activations}
\alias{print.summary.tICA_activations}
\alias{print.tICA_activations}
\title{Summarize a \code{"tICA_activations"} object}
\usage{
\method{summary}{tICA_activations}(object, ...)

\method{print}{summary.tICA_activations}(x, ...)

\method{print}{tICA_activations}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"tICA_activations"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The activations from \code{activations.cifti}}
}
\description{
Summary method for class \code{"tICA_activations"}
}
