% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.nifti.R
\name{templateICA.nifti}
\alias{templateICA.nifti}
\title{Template ICA for NIFTI-format data}
\usage{
templateICA.nifti(
  BOLD_fname,
  template,
  mask_fname = NULL,
  scale = TRUE,
  Q2 = NULL,
  maxQ = NULL,
  maxiter = 100,
  epsilon = 0.001,
  verbose = TRUE,
  write_dir = NULL,
  time_inds = NULL
)
}
\arguments{
\item{BOLD_fname}{File path of NIFTI-format timeseries data (ending in .nii or .nii.gz).}

\item{template}{Result of call to \code{estimate_template.nifti}, either (1) an object of class
\code{template.nifti} OR (2) the file path and basename prefix (e.g., the part before "_mean.nii"
of NIFTI files written out by \code{estimate_template.nifti}.}

\item{mask_fname}{(Optional) File path to brain mask (0/1 or logical). Will also try to use the brain mask
output by \code{estimate_template.nifti} if available.}

\item{scale}{Logical indicating whether BOLD data should be scaled by the spatial
standard deviation before model fitting. If done when estimating templates, should be done here too.}

\item{Q2}{The number of nuisance ICs to identify. If NULL, will be estimated. Only provide \eqn{Q2} or \eqn{maxQ} but not both.}

\item{maxQ}{Maximum number of ICs (template+nuisance) to identify (L <= maxQ <= T). Only provide \eqn{Q2} or \eqn{maxQ} but not both.}

\item{maxiter}{Maximum number of EM iterations. Default: 100.}

\item{epsilon}{Smallest proportion change between iterations. Default: 0.001.}

\item{verbose}{If \code{TRUE} (default), display progress of algorithm.}

\item{write_dir}{Where should any output files be written? \code{NULL} (default) will write them to the current working directory.}

\item{time_inds}{Subset of fMRI BOLD volumes to include in analysis. If NULL (default), use all volumes.}
}
\value{
A list containing the subject IC estimates (class 'xifti'), the subject IC variance estimates (class 'xifti'), and the result of the model call to \code{templateICA} (class 'dICA')
}
\description{
Run template ICA based on NIFTI-format BOLD data and NIFTI-based template
}
