% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICAr-package.R
\docType{package}
\name{templateICAr-package}
\alias{templateICAr}
\alias{templateICAr-package}
\title{templateICAr: Estimate Brain Networks and Connectivity with ICA and Empirical Priors}
\description{
Implements the template ICA (independent components analysis) model proposed in Mejia et al. (2020) \doi{10.1080/01621459.2019.1679638} and the spatial template ICA model proposed in proposed in Mejia et al. (2022) \doi{10.1080/10618600.2022.2104289}. Both models estimate subject-level brain as deviations from known population-level networks, which are estimated using standard ICA algorithms. Both models employ an expectation-maximization algorithm for estimation of the latent brain networks and unknown model parameters. Includes direct support for 'CIFTI', 'GIFTI', and 'NIFTI' neuroimaging file formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mandymejia/templateICAr}
  \item Report bugs at \url{https://github.com/mandymejia/templateICAr/issues}
}

}
\author{
\strong{Maintainer}: Amanda Mejia \email{mandy.mejia@gmail.com}

Authors:
\itemize{
  \item Damon Pham \email{damondpham@gmail.com} (\href{https://orcid.org/0000-0001-7563-4727}{ORCID})
}

Other contributors:
\itemize{
  \item Daniel Spencer \email{danieladamspencer@gmail.com} (\href{https://orcid.org/0000-0002-9705-3605}{ORCID}) [contributor]
  \item Mary Beth Nebel \email{Nebel@kennedykrieger.org} [contributor]
}

}
