% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{update_A}
\alias{update_A}
\title{Update A for VB FC Template ICA}
\usage{
update_A(mu_tau2, mu_S, cov_S, mu_G, mu_alpha, BOLD, ntime, nICs, nvox)
}
\arguments{
\item{mu_tau2, mu_S, cov_S, mu_G, mu_alpha}{Most recent estimates of posterior
moments for these variables.}

\item{BOLD}{(\eqn{T \times V} matrix) preprocessed fMRI data.}

\item{ntime, nICs, nvox}{Number of timepoints in data, number of ICs, and
the number of data locations.}
}
\value{
List of length two: \code{mu_A} and \code{cov_A}.
}
\description{
Update A for VB FC Template ICA
}
\keyword{internal}
