% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{update_A_Chol}
\alias{update_A_Chol}
\title{Update A for VB FC Template ICA using Cholesky prior for FC}
\usage{
update_A_Chol(
  mu_tau2,
  mu_S,
  E_SSt,
  BOLD,
  ntime,
  nICs,
  template_FC,
  final = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{mu_tau2, mu_S, E_SSt}{Most recent estimates of posterior
moments for these variables.}

\item{BOLD}{(\eqn{T \times V} matrix) preprocessed fMRI data.}

\item{ntime, nICs}{Number of timepoints, number of ICs}

\item{template_FC}{IW parameters for FC template}

\item{final}{If TRUE, return cov_A. Default: \code{FALSE}}

\item{verbose}{If \code{TRUE}, display progress of algorithm. Default: \code{FALSE}.}
}
\value{
List of length two: \code{mu_A} (TxQ) and \code{E_AtA} (QxQ).
}
\description{
Update A for VB FC Template ICA using Cholesky prior for FC
}
\keyword{internal}
