% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{update_G_IW}
\alias{update_G_IW}
\title{Update G for VB FC Template ICA}
\usage{
update_G_IW(mu_A, cov_A, template_FC, ntime, nICs)
}
\arguments{
\item{mu_A, cov_A}{Most recent estimates of posterior
moments for these variables.}

\item{template_FC}{(list) Parameters of functional connectivity template.}

\item{ntime, nICs}{Number of timepoints in data and number of ICs.}
}
\value{
List of length two: \code{nu1} and \code{psi1}.
}
\description{
Update G for VB FC Template ICA
}
\keyword{internal}
