% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{update_S}
\alias{update_S}
\title{Update S for VB FC Template ICA}
\usage{
update_S(mu_tau2, mu_A, cov_A, D_inv_S, BOLD, ntime, nICs, nvox)
}
\arguments{
\item{mu_tau2, mu_A, cov_A, }{Most recent estimates of posterior moments for
these variables.}

\item{ntime, nICs, nvox}{Number of timepoints in data, number of ICs, and
the number of data locations.}

\item{template_mean}{(\eqn{V \times Q} matrix) mean maps for each IC in the
template, where \eqn{Q} is the number of ICs, and \eqn{V=nvox} is the number
of data locations.}

\item{template_var}{(\eqn{V \times Q} matrix) between-subject variance maps
for each IC in the template.
#' @param BOLD (\eqn{T \times V} matrix) preprocessed fMRI data.}
}
\value{
List of length two: \code{mu_A} and \code{cov_A}.
}
\description{
Update S for VB FC Template ICA
}
\keyword{internal}
