% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateTheta.R
\name{UpdateTheta}
\alias{UpdateTheta}
\alias{UpdateTheta.spatial}
\alias{UpdateTheta.independent}
\title{Parameter Estimates in EM Algorithm for Template ICA Model}
\usage{
UpdateTheta.spatial(
  template_mean,
  template_var,
  spde,
  BOLD,
  theta,
  C_diag,
  Hinv,
  common_smoothness = TRUE,
  verbose = FALSE,
  return_kappa_fun = FALSE,
  return_MAP = FALSE,
  dim_reduce_flag
)

UpdateTheta.independent(template_mean, template_var, BOLD, theta, C_diag)
}
\arguments{
\item{template_mean}{mean maps for each IC in template}

\item{template_var}{between-subject variance maps for each IC in template}

\item{spde}{NULL for spatial independence model, otherwise SPDE object representing spatial prior on deviations.}

\item{BOLD}{dimension-reduced fMRI data}

\item{theta}{A list of current parameter estimates (mixing matrix A, noise variance nu0_sq and (for spatial model) SPDE parameters kappa)}

\item{C_diag}{(Qx1) diagonal elements of matrix proportional to residual variance.}

\item{common_smoothness}{If TRUE, use the common smoothness version of the spatial template ICA model, which assumes that all IC's have the same smoothness parameter, \eqn{\kappa}}

\item{verbose}{If TRUE, print progress updates for slow steps.}

\item{return_kappa_fun}{If TRUE, return the log likelihood as a function of kappa in a neighborhood of the MLE (common smoothness model only)}

\item{return_MAP}{If TRUE, returns the posterior mean and precision of the latent fields instead of the parameter estimates}

\item{dim_reduce_flag}{If FALSE, data is in the original resolution (no dimension reduction).}
}
\value{
An updated list of parameter estimates, theta, OR if return_MAP=TRUE, the posterior mean and precision of the latent fields
}
\description{
Parameter Estimates in EM Algorithm for Template ICA Model
}
