% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_diagnosticICA.R
\name{UpdateTheta_diagnosticICA}
\alias{UpdateTheta_diagnosticICA}
\alias{UpdateTheta_diagnosticICA.spatial}
\alias{UpdateTheta_diagnosticICA.independent}
\title{Parameter Estimates in EM Algorithm for Diagnostic ICA Model}
\usage{
UpdateTheta_diagnosticICA.spatial(
  template_mean,
  template_var,
  meshes,
  BOLD,
  theta,
  C_diag,
  s0_vec_list,
  D_list,
  Dinv_s0_list,
  verbose = FALSE,
  return_MAP = FALSE,
  update = c("all", "kappa", "A"),
  ignore_determinant = TRUE
)

UpdateTheta_diagnosticICA.independent(
  template_mean,
  template_var,
  template_var_max,
  BOLD,
  theta,
  C_diag,
  return_MAP = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{template_mean}{(A list of G matrices, each VxL) template mean estimates for each group 1 to G}

\item{template_var}{(A list of G matrices, each VxL) template variance estimates for each group 1 to G}

\item{meshes}{Either NULL (assume spatial independence) or a list of objects of type \code{templateICA_mesh} created by \code{make_mesh()} (each list element corresponds to one brain structure)}

\item{BOLD}{(VxL matrix) dimension-reduced fMRI data}

\item{theta}{(list) current parameter estimates}

\item{C_diag}{(Lx1) diagonal elements of residual covariance after dimension reduction}

\item{s0_vec_list}{List of vectorized template means (one for each group 1 to G)}

\item{D_list}{List of sparse diagonal matrices of template standard deviations (one for each group 1 to G)}

\item{Dinv_s0_list}{List of D^{-1} times s0_vec (one for each group 1 to G)}

\item{verbose}{If TRUE, display progress of algorithm}

\item{return_MAP}{If TRUE, return the posterior mean and precision of the latent fields and group membership probabilities instead of the parameter estimates}

\item{update}{Which parameters to update. Either "all", "A" or "kappa".}

\item{ignore_determinant}{For spatial model only. If TRUE, ignore the normalizing constant in p(y|z) when computing posterior probabilities of z.}

\item{template_var_max}{The maximum of the template variance across group}
}
\value{
An updated list of parameter estimates, theta, OR if return_MAP=TRUE, the posterior mean and precision of the latent fields and posterior probabilities of group membership
}
\description{
Parameter Estimates in EM Algorithm for Diagnostic ICA Model
}
