% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_FCtemplateICA.R
\name{UpdateTheta_FCtemplateICA}
\alias{UpdateTheta_FCtemplateICA}
\title{Update FC Template ICA parameters (tau_sq, alpha, G)}
\usage{
UpdateTheta_FCtemplateICA(
  template_mean,
  template_var,
  template_FC,
  prior_params,
  BOLD,
  post_sums,
  verbose
)
}
\arguments{
\item{template_mean}{(\eqn{V \times Q} matrix) mean maps for each IC in template}

\item{template_var}{(\eqn{V \times Q} matrix) between-subject variance maps for each IC in template}

\item{template_FC}{(list) Parameters of functional connectivity template}

\item{prior_params}{Alpha and beta parameters of IG prior on tau^2 (error variance)}

\item{BOLD}{(\eqn{V \times Q} matrix) dimension-reduced fMRI data}

\item{post_sums}{TBD}

\item{verbose}{If \code{TRUE}, display progress of algorithm. Default: \code{FALSE}.}
}
\description{
Update FC Template ICA parameters (tau_sq, alpha, G)
}
