% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_diagnosticICA.R
\name{UpdateThetaSQUAREM_diagnosticICA}
\alias{UpdateThetaSQUAREM_diagnosticICA}
\title{Helper function for SQUAREM for estimating parameters}
\usage{
UpdateThetaSQUAREM_diagnosticICA(
  theta_vec,
  template_mean,
  template_var,
  meshes,
  BOLD,
  C_diag,
  s0_vec_list,
  D_list,
  Dinv_s0_list,
  verbose
)
}
\arguments{
\item{theta_vec}{Vector of initial parameter values}

\item{template_mean}{Passed to UpdateTheta_diagnosticICA function}

\item{template_var}{Passed to UpdateTheta_diagnosticICA function}

\item{meshes}{Passed to UpdateTheta_diagnosticICA function}

\item{BOLD}{Passed to UpdateTheta_diagnosticICA function}

\item{C_diag}{Passed to UpdateTheta_diagnosticICA function}

\item{s0_vec_list}{Passed to UpdateTheta_diagnosticICA function}

\item{D_list}{Passed to UpdateTheta_diagnosticICA function}

\item{Dinv_s0_list}{Passed to UpdateTheta_diagnosticICA function}

\item{verbose}{Passed to UpdateTheta_diagnosticICA function}
}
\value{
Vector of updated parameter values
}
\description{
Helper function for SQUAREM for estimating parameters
}
