% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_templateICA.R
\name{UpdateThetaSQUAREM_templateICA}
\alias{UpdateThetaSQUAREM_templateICA}
\title{Helper function for SQUAREM for estimating parameters}
\usage{
UpdateThetaSQUAREM_templateICA(
  theta_vec,
  template_mean,
  template_var,
  mesh,
  BOLD,
  C_diag,
  s0_vec,
  D,
  Dinv_s0,
  common_smoothness,
  verbose
)
}
\arguments{
\item{theta_vec}{Vector of initial parameter values}

\item{template_mean}{Passed to UpdateTheta_templateICA function}

\item{template_var}{Passed to UpdateTheta_templateICA function}

\item{mesh}{Passed to UpdateTheta_templateICA function}

\item{BOLD}{Passed to UpdateTheta_templateICA function}

\item{C_diag}{Passed to UpdateTheta_templateICA function}

\item{s0_vec}{Passed to UpdateTheta_templateICA function}

\item{D}{Passed to UpdateTheta_templateICA function}

\item{Dinv_s0}{Passed to UpdateTheta_templateICA function}

\item{common_smoothness}{Passed to UpdateTheta_templateICA function}

\item{verbose}{Passed to UpdateTheta_templateICA function}
}
\value{
Vector of updated parameter values
}
\description{
Helper function for SQUAREM for estimating parameters
}
