% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{UT2mat}
\alias{UT2mat}
\title{Transform upper-triangular elements to matrix form}
\usage{
UT2mat(x, diag = TRUE)
}
\arguments{
\item{x}{Vector of upper triangular values}

\item{diag}{Are diagonal values included in x?  Default: \code{TRUE}.}
}
\description{
Transform upper-triangular elements to matrix form
}
