% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMRItools_devTemp.R
\name{var_decomp}
\alias{var_decomp}
\title{Compute variance decomposition}
\usage{
var_decomp(x, verbose = FALSE)
}
\arguments{
\item{x}{measurements by subjects by variables}

\item{verbose}{If \code{TRUE}, display progress of algorithm}
}
\value{
The variance decomposition
}
\description{
Compute variance decomposition
}
\keyword{internal}
