% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.utils.R
\name{var_sq_err_constrained}
\alias{var_sq_err_constrained}
\title{Compute the overall error between empirical and theoretical variance of CORRELATION matrix elements}
\usage{
var_sq_err_constrained(nu, p, var, xbar, M = 10000)
}
\arguments{
\item{nu}{Inverse Wishart degrees of freedom parameter}

\item{p}{Matrix dimension for IW distribution}

\item{var}{Empirical between-subject variances of CORRELATION matrix (upper triangle)}

\item{xbar}{Empirical mean of CORRELATION matrix (upper triangle)}

\item{M}{Penalty to assign if theoretical variance is smaller than empirical variance}
}
\value{
Sum of squared difference between the empirical and theoretical IW variance of CORRELATION matrix,
but with constraint that theoretical variances must not be smaller than empirical variances
}
\description{
Compute the overall error between empirical and theoretical variance of CORRELATION matrix elements
}
