% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{VB_FCtemplateICA}
\alias{VB_FCtemplateICA}
\title{VB Algorithm for FC Template ICA Model}
\usage{
VB_FCtemplateICA(
  template_mean,
  template_var,
  template_FC,
  prior_params = c(0.001, 0.001),
  BOLD,
  A0,
  S0,
  S0_var,
  maxiter = 100,
  epsilon = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{template_mean}{(\eqn{V \times Q} matrix) mean maps for each IC in template,
where \eqn{Q} is the number of ICs, \eqn{V=nvox} is the number of data locations.}

\item{template_var}{(\eqn{V \times Q} matrix) between-subject variance maps for each IC in template}

\item{template_FC}{(list) Parameters of functional connectivity template}

\item{prior_params}{Alpha and beta parameters of IG prior on tau^2 (error variance)}

\item{BOLD}{(\eqn{V \times T} matrix) preprocessed fMRI data}

\item{maxiter}{Maximum number of VB iterations. Default: 100.}

\item{epsilon}{Smallest proportion change in parameter estimates between iterations. Default: 0.01.}

\item{verbose}{If \code{TRUE}, display progress of algorithm. Default: \code{FALSE}.}

\item{AS_0}{(list) initial guess at latent variables: A (\eqn{TxQ} mixing matrix),
and S (\eqn{QxV} matrix of spatial ICs)}
}
\description{
VB Algorithm for FC Template ICA Model
}
