% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.R
\name{activations}
\alias{activations}
\title{Identify areas of activation in each independent component map}
\usage{
activations(
  result,
  u = 0,
  alpha = 0.1,
  type = ">",
  method_p = "BH",
  verbose = FALSE
)
}
\arguments{
\item{result}{Fitted stICA or tICA model object (of class stICA or tICA)}

\item{u}{Activation threshold, default = 0}

\item{alpha}{Significance level for joint PPM, default = 0.1}

\item{type}{Type of region.  Default is '>' (positive excursion region).}

\item{method_p}{If result is type tICA, the type of multiple comparisons correction to use for p-values, or NULL for no correction.  See \code{help(p.adjust)}.}

\item{verbose}{If TRUE, display progress of algorithm}
}
\value{
A list containing activation maps for each IC and the joint and marginal PPMs for each IC.
}
\description{
Identify areas of activation in each independent component map
}
