% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{bdiag_m2}
\alias{bdiag_m2}
\title{Bdiag m2}
\usage{
bdiag_m2(mat, N)
}
\arguments{
\item{mat}{a k x k 'matrix'}

\item{N}{how many times to repeat \code{mat}}
}
\value{
a sparse (N\emph{k x N}k) matrix of class \code{"dgCMatrix"}.
}
\description{
Bdiag m2
}
