% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIGP.R
\name{datProcFUN.cifti}
\alias{datProcFUN.cifti}
\title{CIFTI data processing function for MIGP}
\usage{
datProcFUN.cifti(
  dat,
  brainstructures = c("left", "right"),
  resamp_res = NULL,
  center_Bcols = FALSE,
  scale = TRUE,
  detrend_DCT = 0
)
}
\description{
CIFTI data processing function for MIGP
}
