% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticICA.nifti.R
\name{diagnosticICA.nifti}
\alias{diagnosticICA.nifti}
\title{Run diagnostic ICA based on NIFTI-format BOLD data and NIFTI-based template}
\usage{
diagnosticICA.nifti(
  nifti_fname,
  templates,
  scale = TRUE,
  maxQ = 100,
  maxiter = 100,
  epsilon = 0.01,
  verbose = TRUE,
  out_fname = NULL
)
}
\arguments{
\item{nifti_fname}{File path of NIFTI BOLD timeseries data}

\item{templates}{Set of templates, each the result of call to \code{estimate_template.nifti()}
(one template for each group). A list of objects of class \code{template.nifti}.}

\item{scale}{Logical indicating whether BOLD data should be scaled by the spatial
standard deviation before model fitting. If done when estimating templates, should be done here too.}

\item{maxQ}{Maximum number of ICs (template+nuisance) to identify (L <= maxQ <= T)}

\item{maxiter}{Maximum number of EM iterations}

\item{epsilon}{Smallest proportion change between iterations (e.g. .01)}

\item{verbose}{If TRUE, display progress of algorithm}

\item{out_fname}{The path and base name prefix of the NIFTI files to write.
Will be appended with "_subjICmean.nii" for IC mean maps and "_subjICvar.nii" for
IC variance maps.}
}
\value{
A list containing the subject IC estimates (class 'xifti'), the subject IC variance estimates (class 'xifti'), and the result of the model call to \code{diagnosticICA} (class 'dICA')
}
\description{
Run diagnostic ICA based on NIFTI-format BOLD data and NIFTI-based template
}
