% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticICA.R
\name{diagnosticICA}
\alias{diagnosticICA}
\title{Diagnostic ICA}
\usage{
diagnosticICA(
  template_mean,
  template_var,
  BOLD,
  scale = TRUE,
  meshes = NULL,
  Q2 = NULL,
  maxQ = NULL,
  maxiter = 100,
  epsilon = 0.01,
  verbose = TRUE,
  kappa_init = 0.5
)
}
\arguments{
\item{template_mean}{(A list of G matrices, each VxL) template mean estimates for each group 1 to G}

\item{template_var}{(A list of G matrices, each VxL) template variance estimates for each group 1 to G}

\item{BOLD}{(VxT matrix) BOLD fMRI data matrix, where T is the number of volumes (time points) and V is the number of brain locations}

\item{scale}{Logical indicating whether BOLD data should be scaled by the spatial standard deviation before model fitting. If done when estimating templates, should be done here too.}

\item{meshes}{Either NULL (assume spatial independence) or a list of objects of type \code{templateICA_mesh} created by \code{make_mesh()} (each list element corresponds to one brain structure)}

\item{Q2}{The number of nuisance ICs to identify. If NULL, will be estimated. Only provide Q2 or maxQ but not both.}

\item{maxQ}{Maximum number of ICs (template+nuisance) to identify (L <= maxQ <= T). Only provide Q2 or maxQ but not both.}

\item{maxiter}{Maximum number of EM iterations}

\item{epsilon}{Smallest proportion change between iterations (e.g. .01)}

\item{verbose}{If TRUE, display progress of algorithm}

\item{kappa_init}{Starting value for kappa.  If NULL, starting value will be determined automatically.}
}
\value{
A list containing the posterior probabilities of group membership, 
 the estimated independent components S (a VxL matrix), their mixing matrix 
 A (a TxL matrix), the number of nuisance ICs estimated (Q_nuis)
}
\description{
Perform diagnostic independent component analysis (ICA) using expectation-maximization (EM)
}
