% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticICA.R
\name{diagnosticICA}
\alias{diagnosticICA}
\title{Diagnostic ICA}
\usage{
diagnosticICA(
  template_mean,
  template_var,
  BOLD,
  scale = TRUE,
  meshes = NULL,
  Q2 = 0,
  Q2_max = NULL,
  maxiter = 100,
  epsilon = 0.01,
  verbose = TRUE,
  kappa_init = 0.2
)
}
\arguments{
\item{template_mean}{(A list of G matrices, each \eqn{VxL}) template mean
estimates for each group \eqn{1} to \eqn{G}.}

\item{template_var}{(A list of G matrices, each \eqn{VxL}) template variance
estimates for each group \eqn{1} to \eqn{G}.}

\item{BOLD}{(\eqn{V \times T} matrix) BOLD fMRI data matrix, where \eqn{T} is the number
of volumes (time points) and \eqn{V} is the number of brain locations}

\item{scale}{Should BOLD data be scaled by the spatial standard deviation
before model fitting? Default: \code{TRUE}. If done when estimating
templates, should be done here too.}

\item{meshes}{Either \code{NULL} (assume spatial independence) or a list of objects
of type \code{templateICA_mesh} created by \code{make_mesh} (each list element
corresponds to one brain structure)}

\item{Q2, Q2_max}{Obtain dual regression estimates after denoising? Denoising is
based on modeling and removing nuisance ICs. It may result in a cleaner
estimate for smaller datasets, but it may be unnecessary (and time-consuming)
for larger datasets.

Set \code{Q2} to control denoising: use a positive integer to specify the
number of nuisance ICs, \code{NULL} to have the number of nuisance ICs
estimated by PESEL, or zero (default) to skip denoising.

If \code{is.null(Q2)}, use \code{Q2_max} to specify the maximum number of
nuisance ICs that should be estimated by PESEL. \code{Q2_max} must be less
than \eqn{T * .75 - Q} where \eqn{T} is the number of timepoints in each
fMRI scan and \eqn{Q} is the number of group ICs. If \code{NULL} (default),
\code{Q2_max} will be set to \eqn{T * .50 - Q}, rounded.}

\item{maxiter}{Maximum number of EM iterations. Default: 100.}

\item{epsilon}{Smallest proportion change between iterations. Default: 0.01.}

\item{verbose}{If \code{TRUE} (default), display progress of algorithm.}

\item{kappa_init}{Starting value for kappa.  If \code{NULL}, starting value
will be determined automatically. Default: 0.4.}
}
\value{
A list containing the posterior probabilities of group membership,
the estimated independent components \strong{S} (a \eqn{VxL} matrix), their mixing matrix
\strong{A} (a \eqn{TxL} matrix), the number of nuisance ICs estimated (Q_nuis)
}
\description{
Perform diagnostic independent component analysis (ICA) using expectation-maximization (EM)
}
\keyword{internal}
