% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_reg.R
\name{dual_reg}
\alias{dual_reg}
\title{Dual Regression}
\usage{
dual_reg(dat, GICA, scale = FALSE)
}
\arguments{
\item{dat}{Subject-level fMRI data (\eqn{VxT})}

\item{GICA}{Group-level independent components (\eqn{VxQ})}

\item{scale}{A logical value indicating whether the fMRI timeseries should 
be scaled by the image standard deviation.}
}
\value{
A list containing the subject-level independent components \strong{S} (\eqn{VxQ}), 
 subject-level mixing matrix \strong{A} (\eqn{TxQ}), and the row- and column- centered fMRI 
 data (\eqn{VxT}).
}
\description{
Dual Regression
}
