% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_reg.R
\name{dual_reg}
\alias{dual_reg}
\title{Dual Regression}
\usage{
dual_reg(dat, GICA)
}
\arguments{
\item{dat}{Subject-level fMRI data (TxV)}

\item{GICA}{Group-level independent components (QxV)}
}
\value{
A list containing the subject-level independent components S (QxV), subject-level mixing matrix A (TxQ), and the row- and column- centered fMRI data (TxV)
}
\description{
Dual Regression
}
