% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_reg.R
\name{dual_reg}
\alias{dual_reg}
\title{Dual Regression}
\usage{
dual_reg(dat, GICA, center = FALSE, scale = FALSE, normA = FALSE)
}
\arguments{
\item{dat}{Subject-level fMRI data (\eqn{VxT})}

\item{GICA}{Group-level independent components (\eqn{VxQ})}

\item{center, scale}{A logical value indicating whether the fMRI timeseries should
be centered and/or scaled. See \link{\code{scale_BOLD}}.}

\item{normA}{Normalize the A matrix (spatial maps)?}
}
\value{
A list containing the subject-level independent components \strong{S} (\eqn{VxQ}), 
 and subject-level mixing matrix \strong{A} (\eqn{TxQ}).
}
\description{
Dual Regression
}
