% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_reg.R
\name{dual_reg2}
\alias{dual_reg2}
\title{Dual Regression wrapper}
\usage{
dual_reg2(
  BOLD,
  BOLD2 = NULL,
  GICA,
  scale = TRUE,
  normA = FALSE,
  format = c("infer", "CIFTI", "xifti", "NIFTI", "nifti", "data"),
  expectedV = NULL,
  brainstructures = c("left", "right"),
  mask = NULL,
  Q2 = 0,
  maxQ = NULL,
  verbose = TRUE
)
}
\arguments{
\item{BOLD}{subject-level fMRI timeseries data in one of the following formats: 
 a CIFTI file path, a \code{"xifti"}, a NIFTI file path, a \code{"nifti"}, or
 a \eqn{V \times T} numeric matrix.

 If \code{BOLD2} is provided it must be in the same format as \code{BOLD}; 
 \code{BOLD} will be the test data and \code{BOLD2} will be the retest data. 
 If \code{BOLD2} is not provided, \code{BOLD} will be split in half; 
 the first half will be the test data and the second half will be the retest data.}

\item{GICA}{Group ICA maps in a format compatible with \code{BOLD}. Can also be
vectorized (a \eqn{V \times T} numeric matrix) no matter the format of \code{BOLD}.}

\item{scale}{Scale \code{BOLD} (and \code{BOLD2}) by its mean spatial
standard deviation before computing dual regression? Default: \code{TRUE}.}

\item{normA}{Normalize the A matrix (spatial maps)?}

\item{format}{Expected format of \code{BOLD} and \code{BOLD2}. Should be one of the following:
\code{"infer"} (default), a \code{"CIFTI"} file, a \code{"xifti"} object, a
\code{"NIFTI"} file, a \code{"nifti"} object, or a \code{"data"} matrix.}

\item{expectedV}{The number of rows to expect in the data. Or if the fMRI data
is a NIFTI or "nifti", this is the number of in-mask locations.}

\item{brainstructures}{Only applies if \code{format} is \code{"CIFTI"}. 
Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("left","right")} (cortical surface only).}

\item{mask}{Required if and only if the entries of \code{BOLD} are NIFTI file paths or
\code{"nifti"} objects. A binary brain map of the same size as the fMRI data, with
\code{1} corresponding to in-mask voxels.}

\item{Q2, maxQ}{Denoise the dual regression estimate? Denoising is based on modeling and
removing nuisance ICs. It may result in a cleaner estimate, but will take longer to compute. 
If both are \code{NULL},
denoising will be performed, with the number of nuisance ICs estimated for \code{BOLD} and \code{BOLD2}
separately. If \code{Q2==0}, do not denoise (default). Otherwise, specify one or the other:
use \code{Q2} to specify the number of nuisance ICs, or \code{maxQ} to specify the number of
total ICs (template/group + nuisance). \eqn{L <= (L+Q2) = maxQ <= T}, where \eqn{L} is the number
of template ICs and \eqn{T} is the number of timepoints in each fMRI scan.}

\item{verbose}{Display progress updates? Default: \code{TRUE}.}
}
\description{
Wrapper to \code{dual_reg}. Handles NIFTI or CIFTI file input, halves the 
data in absense of retest data, and may remove nuisance regressors.
}
\keyword{internal}
