% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_reg.R
\name{dual_reg2}
\alias{dual_reg2}
\title{Dual Regression wrapper}
\usage{
dual_reg2(
  BOLD,
  BOLD2 = NULL,
  GICA,
  center_rows = TRUE,
  center_cols = TRUE,
  scale = TRUE,
  detrend_DCT = 0,
  center_Gcols = TRUE,
  normA = FALSE,
  format = c("infer", "CIFTI", "xifti", "NIFTI", "nifti", "data"),
  brainstructures = c("left", "right"),
  mask = NULL,
  Q2 = 0,
  maxQ = NULL,
  verbose = TRUE
)
}
\arguments{
\item{BOLD, BOLD2}{Subject-level fMRI data in one of the following formats: 
 a CIFTI file path, a \code{"xifti"}, a NIFTI file path, a \code{"nifti"}, or
 a \eqn{V \times T} numeric matrix.

 If \code{BOLD2} is provided it must be in the same format as \code{BOLD}; 
 \code{BOLD} will be the test data and \code{BOLD2} will be the retest data. 
 If \code{BOLD2} is not provided, \code{BOLD} will be split in half; 
 the first half will be the test data and the second half will be the retest data.}

\item{GICA}{Group ICA maps in a format compatible with \code{BOLD}. Can also be a
(vectorized) numeric matrix (\eqn{V \times Q}) no matter the format of \code{BOLD}.}

\item{center_rows, center_cols}{Center BOLD data across rows (each data location's time series) or columns (each time point's image)? Default: \code{TRUE} for both.}

\item{scale}{A logical value indicating whether the fMRI timeseries should be scaled by the image standard deviation.}

\item{detrend_DCT}{Detrend the data? This is the number of DCT bases to use for detrending. If \code{0} (default), do not detrend.}

\item{center_Gcols}{Center GICA across columns (each ICA)? Default: \code{TRUE}.}

\item{normA}{Scale each IC timeseries (column of \eqn{A}) in the dual regression 
estimates? Default: \code{FALSE}. (The opposite scaling will be applied to \eqn{S}
such that the product \eqn{A \times S} remains the same).}

\item{format}{Expected format of \code{BOLD} and \code{BOLD2}. Should be one of the following:
\code{"infer"} (default), a \code{"CIFTI"} file path, a \code{"xifti"} object, a
\code{"NIFTI"} file path, a \code{"nifti"} object, or a \code{"data"} matrix.}

\item{brainstructures}{Only applies if \code{format} is \code{"CIFTI"}. 
Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("left","right")} (cortical surface only).}

\item{mask}{Required if and only if \code{BOLD} is a NIFTI file path or
\code{"nifti"} object. This is a brain map formatted as a binary array of the same 
size as the fMRI data, with \code{TRUE} corresponding to in-mask voxels.}

\item{Q2, maxQ}{Obtain dual regression estimates after denoising? Denoising is based on modeling and
removing nuisance ICs. It may result in a cleaner estimate for smaller datasets, but it may be unnecessary (and time-consuming) for larger datasets.
If both arguments are \code{NULL}, denoising will be performed, with the number of nuisance 
ICs estimated for \code{BOLD} and \code{BOLD2} separately. Otherwise, specify one or the other:
use \code{Q2} to specify the number of nuisance ICs, or \code{maxQ} to specify the number of
total ICs (group + nuisance, or \eqn{Q + Q2}). Set either to zero to skip denoising.
Default: \code{Q2==0} (do not denoise).

The valid inputs are \eqn{Q <= (Q+Q2) = maxQ <= T}, where \eqn{Q} is the number
of group ICs and \eqn{T} is the number of timepoints in each fMRI scan.}

\item{verbose}{Display progress updates? Default: \code{TRUE}.}
}
\description{
Wrapper to \code{dual_reg}. Can handle NIFTI or CIFTI file input, halving the 
data in absense of retest data, and removing nuisance regressors.
}
\details{
Used to compute dual regression for the purpose of template estimation, and 
 to obtain initial subject-level IC estimates for template ICA. In the latter
 case, \code{GICA} should be the template mean.
}
\keyword{internal}
