% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_reg.R
\name{dual_reg2}
\alias{dual_reg2}
\title{Dual Regression wrapper}
\usage{
dual_reg2(
  BOLD,
  BOLD2 = NULL,
  GICA,
  scale = FALSE,
  format = c("infer", "data", "CIFTI", "NIFTI"),
  Q2 = NULL,
  maxQ = NULL,
  dim_expect = NULL,
  brainstructures = c("left", "right"),
  verbose = FALSE
)
}
\arguments{
\item{BOLD, BOLD2}{subject-level BOLD data: a CIFTI file, NIFTI file, or data matrix.
 Cannot be a \code{"xifti"} object, \code{"nifti"} object, etc. If \code{BOLD2} is provided 
 it must be in the same format as \code{BOLD}; \code{BOLD} will be the test data and
 \code{BOLD2} will be the retest data. If \code{BOLD2} is not provided,
 \code{BOLD} will be split in half; the first half will be the test data and
 the second half will be the retest data.

 NIFTI implementation is not complete yet (will return error).}

\item{GICA}{Group ICA data matrix. Cannot be a \code{"xifti"} object, \code{"nifti"} object, file, etc.}

\item{scale}{Logical indicating whether BOLD data should be scaled by the
spatial standard deviation before template estimation.}

\item{format}{Expected format of \code{BOLD} and \code{BOLD2}. Should be one of the following:
\code{"infer"} (default), a \code{"data"} matrix, \code{"CIFTI"} file, or \code{"NIFTI"} file.}

\item{Q2}{The number of nuisance ICs to identify. If \code{NULL}, will be estimated.
Only provide \code{Q2} or \code{maxQ} but not both.}

\item{maxQ}{Maximum number of ICs (template+nuisance) to identify 
(L <= maxQ <= T). If \code{maxQ == L}, then do not remove any nuisance regressors.
Only provide \code{Q2} or \code{maxQ} but not both.}

\item{dim_expect}{If \code{!(format == "NIFTI")}, this is the number of rows to expect in the data.
Otherwise, ...}

\item{brainstructures}{Only applies if \code{format} is \code{"CIFTI"}. 
Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("left","right")} (cortical surface only).}

\item{verbose}{If \code{TRUE}, display progress updates}
}
\description{
Wrapper to \code{dual_reg}. Handles NIFTI or CIFTI file input, halves the 
data in absense of retest data, and may remove nuisance regressors.
}
\keyword{internal}
