% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{ELBO}
\alias{ELBO}
\title{ELBO (Evidence Lower Bound)}
\usage{
ELBO(mu_S, cov_S, cov_A, cov_alpha, template_mean, template_var, ntime)
}
\arguments{
\item{mu_S}{(Q x V)}

\item{cov_S}{(Q x Q x V)}

\item{cov_A}{(Q x Q) (common across all a_t)}

\item{cov_alpha}{(Q x Q)}

\item{template_mean}{(VxQ) The template mean for S}

\item{template_var}{(VxQ) The template variance for S}

\item{ntime}{Number of timepoints in data}
}
\value{
ELBO
}
\description{
Computes the Evidence Lower Bound for the VB Algorithm for FC Template ICA
}
\keyword{internal}
