% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{ELBO}
\alias{ELBO}
\title{ELBO}
\usage{
ELBO(
  mu_S,
  cov_S,
  mu_A,
  cov_A,
  mu_alpha,
  cov_alpha,
  mu_G,
  psi_G,
  nu_G,
  mu_tau2,
  alpha1,
  beta1,
  template_mean,
  template_var,
  template_FC,
  BOLD,
  ntime,
  nICs,
  nvox,
  prior_params
)
}
\arguments{
\item{mu_S, cov_S}{(nICs, nvox) and (nICs, nICs, nvox)}

\item{mu_A, cov_A}{TxQ, (QxQ) common across T}

\item{mu_alpha, cov_alpha}{Q, QxQ}

\item{mu_G, psi_G, nu_G}{QxQ, QxQ, 1}

\item{mu_tau2, alpha1, beta1}{#V, 1, V}

\item{template_mean, template_var, template_FC}{The templates}

\item{BOLD, }{#TxV (has been transposed)}

\item{ntime, nICs, nvox}{Number of timepoints in data, number of ICs, and
the number of data locations.}

\item{prior_params}{}
}
\value{
ELBO
}
\description{
Does ELBO ...
}
\keyword{internal}
