% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template.cifti}
\alias{estimate_template.cifti}
\title{Estimate CIFTI template}
\usage{
estimate_template.cifti(
  cifti_fnames,
  cifti_fnames2 = NULL,
  GICA_fname,
  inds = NULL,
  scale = TRUE,
  brainstructures = c("left", "right"),
  var_method = c("unbiased", "non-negative"),
  keep_DR = FALSE,
  Q2 = NULL,
  maxQ = NULL,
  verbose = TRUE,
  out_fname = NULL
)
}
\arguments{
\item{cifti_fnames}{Vector of file paths of CIFTI-format fMRI timeseries
(*.dtseries.nii) for template estimation}

\item{cifti_fnames2}{Vector of file paths of "retest" CIFTI-format fMRI
timeseries (*.dtseries.nii) for template estimation.  Must be from the same
subjects and in the same order as cifti_fnames.  If none specified, will
create pseudo test-retest data from single session.}

\item{GICA_fname}{File path of CIFTI-format group ICA maps (ending in .d*.nii)}

\item{inds}{Numeric indices of the group ICs to include in template. If NULL,
use all group ICs.}

\item{scale}{Logical indicating whether BOLD data should be scaled by the
spatial standard deviation before template estimation.}

\item{brainstructures}{Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("left","right")} (cortical surface only).}

\item{var_method}{\code{"unbiased"} (default) or \code{"non-negative"}}

\item{keep_DR}{Keep the DR estimates? If \code{TRUE}, the DR estimates are
returned too. If a single file path, save the DR estimates as an .rds file to
that location. If \code{FALSE} (default), do not save the DR estimates and
only return the templates.}

\item{Q2}{The number of nuisance ICs to identify. If NULL, will be estimated. Only provide \eqn{Q2} or \eqn{maxQ} but not both.}

\item{maxQ}{Maximum number of ICs (template+nuisance) to identify (L <= maxQ <= T). Only provide \eqn{Q2} or \eqn{maxQ} but not both.}

\item{verbose}{If \code{TRUE}, display progress updates}

\item{out_fname}{(Required if templates are to be resampled to a lower spatial
resolution, usually necessary for spatial template ICA.) The path and base name
prefix of the CIFTI files to write. Will be appended with "_mean.dscalar.nii" for
template mean maps and "_var.dscalar.nii" for template variance maps.}
}
\value{
List of two elements: template mean of class xifti and template
 variance of class xifti
}
\description{
Estimate template for Template or Diagnostic ICA based on CIFTI-format data
}
