% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template}
\alias{estimate_template}
\alias{estimate_template.cifti}
\alias{estimate_template.nifti}
\title{Estimate template}
\usage{
estimate_template(
  BOLD,
  BOLD2 = NULL,
  GICA,
  inds = NULL,
  center_rows = TRUE,
  center_cols = TRUE,
  scale = TRUE,
  detrend_DCT = 0,
  center_Gcols = TRUE,
  normA = FALSE,
  Q2 = 0,
  Q2_max = NULL,
  brainstructures = c("left", "right"),
  mask = NULL,
  var_method = c("unbiased", "non-negative"),
  keep_DR = FALSE,
  out_fname = NULL,
  FC = FALSE,
  verbose = TRUE
)

estimate_template.cifti(
  BOLD,
  BOLD2 = NULL,
  GICA,
  inds = NULL,
  center_rows = TRUE,
  center_cols = TRUE,
  scale = TRUE,
  detrend_DCT = 0,
  center_Gcols = TRUE,
  normA = FALSE,
  brainstructures = c("left", "right"),
  var_method = c("unbiased", "non-negative"),
  keep_DR = FALSE,
  Q2 = 0,
  Q2_max = NULL,
  out_fname = NULL,
  FC = FALSE,
  verbose = TRUE
)

estimate_template.nifti(
  BOLD,
  BOLD2 = NULL,
  GICA,
  inds = NULL,
  center_rows = TRUE,
  center_cols = TRUE,
  scale = TRUE,
  detrend_DCT = 0,
  center_Gcols = TRUE,
  normA = FALSE,
  mask = mask,
  var_method = c("unbiased", "non-negative"),
  keep_DR = FALSE,
  Q2 = 0,
  Q2_max = NULL,
  out_fname = NULL,
  FC = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{BOLD, BOLD2}{Vector of subject-level fMRI data in one of the following formats: 
 CIFTI file paths, \code{"xifti"} objects, NIFTI file paths, \code{"nifti"} objects, or
 \eqn{V \times T} numeric matrices, where \eqn{V} is the number of data locations and
 \eqn{T} is the number of timepoints.

 If \code{BOLD2} is provided it must be in the same format as \code{BOLD}; 
 \code{BOLD} will be the test data and \code{BOLD2} will be the retest data. 
 \code{BOLD2} should be the same length as \code{BOLD} and have the same subjects in the same order.
 If \code{BOLD2} is not provided, \code{BOLD} will be split in half; 
 the first half will be the test data and the second half will be the retest data.}

\item{GICA}{Group ICA maps in a format compatible with \code{BOLD}. Can also be a
(vectorized) numeric matrix (\eqn{V \times Q}) no matter the format of \code{BOLD}.}

\item{inds}{Numeric indices of the group ICs to include in the template. If 
 \code{NULL}, use all group ICs (default).

 If \code{inds} is provided, the ICs not included will be removed after calculating
 dual regression, not before. This is because removing the ICs prior to dual 
 regression would leave unmodeled signals in the data, which could bias the 
 templates.}

\item{center_rows, center_cols}{Center BOLD data across rows (each data location's time series) or columns (each time point's image)? Default: \code{TRUE} for both.}

\item{scale}{A logical value indicating whether the fMRI timeseries should be scaled by the image standard deviation.}

\item{detrend_DCT}{Detrend the data? This is the number of DCT bases to use for detrending. If \code{0} (default), do not detrend.}

\item{center_Gcols}{Center GICA across columns (each ICA)? Default: \code{TRUE}.}

\item{normA}{Scale each IC timeseries (column of \eqn{A}) in the dual regression 
estimates? Default: \code{FALSE}. (The opposite scaling will be applied to \eqn{S}
such that the product \eqn{A \times S} remains the same).}

\item{Q2, Q2_max}{Obtain dual regression estimates after denoising? Denoising is
 based on modeling and removing nuisance ICs. It may result in a cleaner 
 estimate for smaller datasets, but it may be unnecessary (and time-consuming)
 for larger datasets. 
 
 Set \code{Q2} to control denoising: use a positive integer to specify the
 number of nuisance ICs, \code{NULL} to have the number of nuisance ICs
 estimated by PESEL, or zero (default) to skip denoising. 

 If \code{is.null(Q2)}, use \code{Q2_max} to specify the maximum number of
 nuisance ICs that should be estimated by PESEL. \code{Q2_max} must be less
 than \eqn{T * .75 - Q} where \eqn{T} is the number of timepoints in each 
 fMRI scan and \eqn{Q} is the number of group ICs. If \code{NULL} (default),
 \code{Q2_max} will be set to \eqn{T * .50 - Q}, rounded.}

\item{brainstructures}{Only applies if the entries of \code{BOLD} are CIFTI file paths. 
Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("left","right")} (cortical surface only).}

\item{mask}{Required if and only if the entries of \code{BOLD} are NIFTI file paths or
\code{"nifti"} objects. This is a brain map formatted as a binary array of the same 
size as the fMRI data, with \code{TRUE} corresponding to in-mask voxels.}

\item{var_method}{Method for estimating the template variance: \code{"unbiased"} 
(default), \code{"non-negative"}, or \code{"both"}. The unbiased template variance is
based on the assumed mixed effects/ANOVA model, whereas the non-negative template
variance adds to it to account for greater potential between-subjects variation.
(The template mean is the same for either choice of \code{var_method}.)}

\item{keep_DR}{Keep the DR estimates? If \code{FALSE} (default), do not save the DR 
estimates and only return the templates. If \code{TRUE}, the DR estimates are
returned too. If a single file path, save the DR estimates as an RDS file at
that location. If a list of two vectors of file paths with the same lengths as
\code{BOLD}, save the DR estimates as individual files at these locations in
the appropriate format (CIFTI, NIFTI, or RDS files, depending on \code{BOLD}).}

\item{out_fname}{Character vector of file path(s) to write the mean and variance templates to.
If one file name is provided, it will be appended with \code{"_mean.dscalar.nii"} for the
template mean map and \code{"_var.dscalar.nii"} for the template variance map. If two
file names are provided, the first will be used for the template mean and the second will
be used for the template variance. If \code{var_method=="both"}, the non-negative template 
variance will be appended with \code{"_var_nn.dscalar.nii"}.}

\item{FC}{Include the functional connectivity template?}

\item{verbose}{Display progress updates? Default: \code{TRUE}.}
}
\value{
A list with two entries, \code{"template_mean"} and \code{"template_var"}. There
 may be more entries too, depending on the function arguments.
}
\description{
Estimate template for Template or Diagnostic ICA based on fMRI data
}
\details{
All fMRI data (entries in \code{BOLD} and \code{BOLD2}, and \code{GICA}) must be in
 the same spatial resolution.
}
