% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template}
\alias{estimate_template}
\title{Estimate template}
\usage{
estimate_template(
  BOLD,
  BOLD2 = NULL,
  GICA,
  inds = NULL,
  scale = TRUE,
  normA = FALSE,
  brainstructures = c("left", "right"),
  mask = NULL,
  var_method = c("unbiased", "non-negative"),
  keep_DR = FALSE,
  Q2 = 0,
  maxQ = NULL,
  out_fname = NULL,
  FC = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{BOLD}{Vector of fMRI timeseries data for template estimation. The data
format can be one of the following: a character vector of CIFTI file paths
(*.dtseries.nii), a list of \code{"xifti"} objects, a character vector of
NIFTI file paths (*.nii), a list of \code{"nifti"} objects, or a list of 
\eqn{V \times T} numeric matrices.}

\item{BOLD2}{Vector of fMRI timeseries retest data for template estimation. Must
be from the same subjects, in the same order, and in the same format as \code{BOLD}.
If \code{NULL} (default), a pseudo test-retest dataset will be made by splitting
each scan of \code{BOLD} in half.}

\item{GICA}{Group ICA maps in a format compatible with \code{BOLD}.}

\item{inds}{Numeric indices of the group ICs to include in the template. If 
 \code{NULL}, use all group ICs (default).

 If \code{inds} is provided, the ICs not included will be removed after calculating
 dual regression, not before. This is because removing the ICs prior to dual 
 regression would leave unmodeled signals in the data, which could bias the 
 templates.}

\item{scale}{Scale each entry of \code{BOLD} and \code{BOLD2} by its mean spatial
standard deviation before computing dual regression? Default: \code{TRUE}.}

\item{normA}{Scale each IC timeseries (column of \eqn{A}) in the dual regression 
estimates? Default: \code{FALSE}. (The opposite scaling will be applied to \eqn{S}
such that the product \eqn{A \times S} remains the same).}

\item{brainstructures}{Only applies if the entries of \code{BOLD} are CIFTI file paths. 
Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("left","right")} (cortical surface only).}

\item{mask}{Required if and only if the entries of \code{BOLD} are NIFTI file paths or
\code{"nifti"} objects. A binary brain map of the same size as the fMRI data, with
\code{1} corresponding to in-mask voxels.}

\item{var_method}{Method for estimating the template variance: \code{"unbiased"} 
(default), \code{"non-negative"}, or \code{"both"}. The unbiased template variance is
based on the assumed mixed effects/ANOVA model, whereas the non-negative template
variance adds to it to account for greater potential between-subjects variation.
(The template mean is the same for either choice of \code{var_method}.)}

\item{keep_DR}{Keep the DR estimates? If \code{FALSE} (default), do not save the DR 
estimates and only return the templates. If \code{TRUE}, the DR estimates are
returned too. If a single file path, save the DR estimates as an .rds file at
that location. If a list of two vectors of file paths with the same lengths as
\code{BOLD}, save the DR estimates as individual files at these locations in
the appropriate format (CIFTI, NIFTI, or .rds, depending on \code{BOLD}).}

\item{Q2, maxQ}{Denoise the dual regression estimates? Denoising is based on modeling and
removing nuisance ICs. It may result in cleaner estimates for smaller datasets, but may
be unnecessary (and time-consuming) for larger datasets. If both are \code{NULL},
denoising will be performed, with the number of nuisance ICs estimated for each fMRI scan
separately. If \code{Q2==0}, do not denoise (default). Otherwise, specify one or the other:
use \code{Q2} to specify the number of nuisance ICs, or \code{maxQ} to specify the number of
total ICs (template/group + nuisance). \eqn{L <= (L+Q2) = maxQ <= T}, where \eqn{L} is the number
of template ICs and \eqn{T} is the number of timepoints in each fMRI scan.}

\item{out_fname}{Character vector of file path(s) to write the mean and variance templates to.
If one file name is provided, it will be appended with \code{"_mean.dscalar.nii"} for the
template mean map and \code{"_var.dscalar.nii"} for the template variance map. If two
file names are provided, the first will be used for the template mean and the second will
be used for the template variance. If \code{var_method=="both"}, the non-negative template 
variance will be appended with \code{"_var_nn.dscalar.nii"}.}

\item{FC}{Include the functional connectivity template?}

\item{verbose}{Display progress updates? Default: \code{TRUE}.}
}
\value{
A list with two entries, \code{"template_mean"} and \code{"template_var"}. There
 may be more entries too, depending on the function arguments.
}
\description{
Estimate template for Template or Diagnostic ICA based on fMRI data
}
\details{
All fMRI data (entries in \code{BOLD} and \code{BOLD2}, and \code{GICA}) must be in
 the same spatial resolution.
}
