% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template_FC}
\alias{estimate_template_FC}
\title{Estimate FC template}
\usage{
estimate_template_FC(FC0, nu_adjust = 1)
}
\arguments{
\item{FC0}{The FC estimates from \code{\link{estimate_template}}.}

\item{nu_adjust}{Factor by which to adjust estimate of nu.  Values < 1 will
inflate the template variance to avoid an over-informative prior on FC.}
}
\description{
Estimate FC template
}
