% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template_from_DR}
\alias{estimate_template_from_DR}
\title{Estimate template from DR estimates}
\usage{
estimate_template_from_DR(DR1, DR2, var_method = c("unbiased", "non-negative"))
}
\arguments{
\item{DR1, DR2}{the test and retest lists of dual regression estimates}

\item{var_method}{\code{"unbiased"} (default) or \code{"non-negative"}}
}
\value{
List of two elements: the mean and variance templates
}
\description{
Estimate template from DR estimates
}
\keyword{internal}
