% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.R
\name{estimate_template_from_DR_two}
\alias{estimate_template_from_DR_two}
\title{Estimate template from DR estimates (when there are two measurements)}
\usage{
estimate_template_from_DR_two(DR1, DR2)
}
\arguments{
\item{DR1, DR2}{the test and retest dual regression estimates (\eqn{N \times L \times V})}
}
\value{
List of two elements: the mean and variance templates
}
\description{
Legacy version of \code{\link{estimate_template_from_DR}}
}
\keyword{internal}
