% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mesh.R
\name{make_mesh_2D}
\alias{make_mesh_2D}
\title{Create INLA mesh and observation weight matrix based on a binary brain mask}
\usage{
make_mesh_2D(mask)
}
\arguments{
\item{mask}{Brain mask (matrix of 0/1 or TRUE/FALSE). Only supply surf OR mask.}
}
\value{
List containing INLA mesh, observation weight matrix (A) for translating between mesh locations and original data locations, the brain mask used to create the mesh, and the number of original and mesh data locations
}
\description{
Create INLA mesh and observation weight matrix based on a binary brain mask
}
\note{
This function requires the \code{INLA} package, which is not a CRAN package. See \url{http://www.r-inla.org/download} for easy installation instructions.
}
