% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_BOLD.R
\name{norm_BOLD}
\alias{norm_BOLD}
\title{Normalize BOLD data}
\usage{
norm_BOLD(
  BOLD,
  center_rows = TRUE,
  center_cols = TRUE,
  scale = FALSE,
  detrend_DCT = 0
)
}
\arguments{
\item{BOLD}{fMRI numeric data matrix (\eqn{V \times T})}

\item{center_rows, center_cols}{Center BOLD data across rows (each data location's time series) or columns (each time point's image)? Default: \code{TRUE} for both.}

\item{scale}{A logical value indicating whether the fMRI timeseries should be scaled by the image standard deviation. Default: \code{FALSE}.}

\item{detrend_DCT}{Detrend the data? This is the number of DCT bases to use for detrending. If \code{0} (default), do not detrend.}
}
\value{
Normalized BOLD data matrix (\eqn{V \times T})
}
\description{
Center the data across space and time, scale, and detrend, in that order.
}
\details{
The BOLD data is centered across time, which removes the mean image by subtracting the mean of each voxel timeseries, and across space,
which removes the global signal.  Both steps are necessary for dual regression, which in the first regression treats time points as observations,
and in the second step treats voxels as observations. Neither regression includes an intercept, so the BOLD timeseries must be centered across
both space and time. The group independent components used in the first regression must also be centered across space.  The mixing matrix estimated
in the first step of dual regression is also centered across time as a result.

In order to ensure that all fMRI data share the same scale, the BOLD data can also be scaled by the global image standard deviation, equal to
\deqn{\sqrt{\frac{1}{T}\sum_{t=1}^T \sigma^2_t}},
where \eqn{\sigma^2_t} is the standard deviation across all voxels at time point \eqn{t}. If scaling is applied to the BOLD timeseries used in
template estimation, it should also be applied to the BOLD timeseries being analyzed with template ICA using the resulting templates to ensure
compatibility of scale.
}
