% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.methods.R
\name{plot.template.cifti}
\alias{plot.template.cifti}
\title{Plot template}
\usage{
\method{plot}{template.cifti}(
  x,
  stat = c("both", "mean", "sd", "var"),
  var_method = c("non-negative", "unbiased"),
  ...
)
}
\arguments{
\item{x}{The template from \code{estimate_template.cifti}}

\item{stat}{\code{"mean"}, \code{"sd"}, or \code{"both"} (default). By
default the square root of the variance template is shown; another option is
\code{stat="var"} to instead display the variance template directly.}

\item{var_method}{\code{"non-negative"} (default) or \code{"unbiased"}}

\item{...}{Additional arguments to \code{view_xifti}}
}
\value{
The plot
}
\description{
Plot template
}
