% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.methods.R
\name{plot.tICA.nifti}
\alias{plot.tICA.nifti}
\title{Plot template}
\usage{
\method{plot}{tICA.nifti}(x, ...)
}
\arguments{
\item{x}{The result of \code{templateICA} with NIFTI data}

\item{...}{Additional arguments}
}
\value{
The plot
}
\description{
Plot template
}
