% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_BOLD.R
\name{scale_BOLD}
\alias{scale_BOLD}
\title{Center BOLD data across space and time and scale by global temporal standard deviation}
\usage{
scale_BOLD(BOLD, scale = FALSE)
}
\arguments{
\item{BOLD}{(TxV matrix) fMRI timeseries data matrix}
}
\value{
Centered and scaled BOLD data (TxV matrix)
}
\description{
Center BOLD data across space and time and scale by global temporal standard deviation
}
