% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_template.R
\name{struct_template}
\alias{struct_template}
\title{Apply data structure to templates}
\usage{
struct_template(
  template,
  FORMAT,
  mask_input,
  params,
  dat_struct,
  GICA_parc_table
)
}
\arguments{
\item{template}{The template}

\item{FORMAT}{"CIFTI", "GIFTI", "NIFTI", or "DATA"}

\item{mask_input}{The input mask}

\item{params}{The params}

\item{dat_struct}{The data structure}
}
\description{
Apply data structure to templates
}
\keyword{internal}
