% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_template.R
\name{struct_template}
\alias{struct_template}
\title{Apply data structure to templates}
\usage{
struct_template(template, FORMAT, dat_struct, params)
}
\arguments{
\item{template}{The template}

\item{FORMAT}{"CIFTI", "GIFTI", "NIFTI", or "DATA"}

\item{dat_struct}{The data structure}

\item{params}{The params}
}
\description{
Apply data structure to templates
}
\keyword{internal}
