% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_template.methods.R
\name{summary.template.gifti}
\alias{summary.template.gifti}
\alias{print.summary.template.gifti}
\alias{print.template.gifti}
\title{Summarize a \code{"template.gifti"} object}
\usage{
\method{summary}{template.gifti}(object, ...)

\method{print}{summary.template.gifti}(x, ...)

\method{print}{template.gifti}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"template.gifti"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The template from \code{estimate_template.gifti}}
}
\value{
A list summarizing the template: data dimensions, options used for
template estimation, etc.

Nothing, invisibly.

Nothing, invisibly.
}
\description{
Summary method for class \code{"template.gifti"}
}
