% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.methods.R
\name{summary.tICA.nifti}
\alias{summary.tICA.nifti}
\alias{print.summary.tICA.nifti}
\alias{print.tICA.nifti}
\title{Summarize a \code{"tICA.nifti"} object}
\usage{
\method{summary}{tICA.nifti}(object, ...)

\method{print}{summary.tICA.nifti}(x, ...)

\method{print}{tICA.nifti}(x, ...)
}
\arguments{
\item{object}{Object of class \code{"tICA.nifti"}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{The template from \code{estimate_template.cifti}}
}
\value{
A list summarizing of the results of the templateICA analysis.

Nothing, invisibly.

Nothing, invisibly.
}
\description{
Summary method for class \code{"tICA.nifti"}
}
