% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.cifti.R
\name{templateICA.cifti}
\alias{templateICA.cifti}
\title{Template ICA for CIFTI}
\usage{
templateICA.cifti(
  cifti_fname,
  template,
  brainstructures = c("left", "right"),
  spatial_model = FALSE,
  surfL_fname = NULL,
  surfR_fname = NULL,
  surfL_fname_vis = NULL,
  surfR_fname_vis = NULL,
  resamp_res = NULL,
  scale = TRUE,
  Q2 = NULL,
  maxQ = NULL,
  maxiter = 100,
  epsilon = 0.001,
  verbose = TRUE,
  kappa_init = 0.2,
  write_dir = NULL,
  rm_mwall = TRUE,
  time_inds = NULL
)
}
\arguments{
\item{cifti_fname}{File path of CIFTI-format timeseries data (ending in .dtseries.nii).}

\item{template}{Result of call to \code{estimate_template.cifti}, either (1) an object of class
\code{template.cifti} OR (2) the file path and basename prefix (the part before "_mean.dscalar.nii"
and "_var.dscalar.nii") of cifti files written out by \code{estimate_template.cifti}. For
spatial template ICA (\code{spatial_model=TRUE}), must be the latter.}

\item{brainstructures}{Character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("left","right")} (cortical surface only).  For spatial
modeling (performed if \code{surfL_fname} and/or \code{surfR_fname} provided), subcortical
locations will be ignored.}

\item{spatial_model}{Should spatial modeling be performed? (Default \eqn{FALSE}) If \code{TRUE}. surface
geometries will be used to fit a spatial Bayesian model. Computational demands (time and memory)
are much higher with spatial_model=TRUE than spatial_model=FALSE, but results are more accurate.}

\item{surfL_fname}{(Required for spatial modeling) File path of GIFTI surface geometry
file representing the left cortex. Only required if \code{brainstructures} includes \code{"left"}.}

\item{surfR_fname}{(Required for spatial modeling) File path of GIFTI surface geometry
file representing the right cortex. Only required if \code{brainstructures} includes \code{"right"}.}

\item{surfL_fname_vis}{(Optional) File path of GIFTI surface geometry file representing the left cortex
to be used for visualization. Only provide if \code{brainstructures} includes \code{"left"}.}

\item{surfR_fname_vis}{(Optional) File path of GIFTI surface geometry file representing the right cortex
to be used for visualization. Only provide if \code{brainstructures} includes \code{"right"}.}

\item{resamp_res}{(Only recommended for spatial modeling) Target resolution for resampling (number of
cortical surface vertices per hemisphere). A value less than 10000 is recommended for computational
feasibility. If \code{NULL} (default) or \code{FALSE}, do not perform resampling.}

\item{scale}{Logical indicating whether BOLD data should be scaled by the spatial
standard deviation before model fitting. If done when estimating templates, should be done here too.}

\item{Q2}{The number of nuisance ICs to identify. If NULL, will be estimated. Only provide \eqn{Q2} or \eqn{maxQ} but not both.}

\item{maxQ}{Maximum number of ICs (template+nuisance) to identify (L <= maxQ <= T). Only provide \eqn{Q2} or \eqn{maxQ} but not both.}

\item{maxiter}{Maximum number of EM iterations. Default: 100.}

\item{epsilon}{Smallest proportion change between iterations. Default: 0.001.}

\item{verbose}{If \code{TRUE} (default), display progress of algorithm.}

\item{kappa_init}{Starting value for kappa.  Default: \code{0.2}.}

\item{write_dir}{Where should any output files be written? \code{NULL} (default) will write them to the current working directory.}

\item{rm_mwall}{Should medial wall (missing data) locations be removed from mesh?  If TRUE, faster estimation but less accurate estimates on boundary of wall.}

\item{time_inds}{Subset of fMRI BOLD volumes to include in analysis. If NULL (default), use all volumes.}
}
\value{
A list containing the subject IC estimates (class 'xifti'), the subject IC variance estimates (class 'xifti'), and the result of the model call to \code{templateICA} (class 'dICA')
}
\description{
Run template ICA based on CIFTI-format BOLD data and CIFTI-based template
}
