% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateICA.R
\name{templateICA}
\alias{templateICA}
\title{Perform template independent component analysis (ICA) using expectation-maximization (EM)}
\usage{
templateICA(template_mean, template_var, BOLD, mesh = NULL,
  maxQ = NULL, maxiter = 100, epsilon = 0.01)
}
\arguments{
\item{template_mean}{(LxV matrix) template mean estimates, i.e. mean of empirical population prior for each of L independent components}

\item{template_var}{(LxV matrix) template variance estimates, i.e. between-subject variance of empirical population prior for each of L ICs}

\item{BOLD}{(TxV matrix) BOLD fMRI data matrix, where T is the number of volumes (time points) and V is the number of brain locations}

\item{mesh}{Either NULL (assume spatial independence) or an object of type \code{templateICA_mesh} created by \code{make_mesh()} (spatial priors are assumed on each independent component)}

\item{maxQ}{Maximum number of ICs (template+nuisance) to identify (L <= maxQ <= T)}

\item{maxiter}{Maximum number of EM iterations}

\item{epsilon}{Smallest proportion change between iterations (e.g. .01)}
}
\value{
A list containing the estimated independent components S (a LxV matrix), their mixing matrix A (a TxL matrix), and the number of nuisance ICs estimated (Q_nuis)
}
\description{
Perform template independent component analysis (ICA) using expectation-maximization (EM)
}
