% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{update_A}
\alias{update_A}
\title{Update A for VB FC Template ICA}
\usage{
update_A(
  mu_tau2,
  mu_S,
  E_SSt,
  BOLD,
  ntime,
  nICs,
  u_samps,
  template_FC,
  final = FALSE
)
}
\arguments{
\item{mu_tau2, mu_S, E_SSt}{Most recent estimates of posterior
moments for these variables.}

\item{BOLD}{(\eqn{T \times V} matrix) preprocessed fMRI data.}

\item{ntime, nICs}{Number of timepoints, number of ICs}

\item{u_samps}{Samples from Gamma for multivariate t prior on A}

\item{template_FC}{IW parameters for FC template}

\item{final}{If TRUE, return cov_A. Default: \code{FALSE}}
}
\value{
List of length two: \code{mu_A} (TxQ) and \code{E_AtA} (QxQ).
}
\description{
Update A for VB FC Template ICA
}
\keyword{internal}
