% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{update_S}
\alias{update_S}
\title{Update S for VB FC Template ICA}
\usage{
update_S(mu_tau2, mu_A, E_AtA, D_inv, D_inv_S, BOLD, nICs, nvox, final = FALSE)
}
\arguments{
\item{mu_tau2, mu_A, E_AtA, }{Most recent posterior estimates}

\item{D_inv, D_inv_S}{Some pre-computed quantities.}

\item{BOLD}{(\eqn{T \times V} matrix) preprocessed fMRI data.}

\item{nICs, nvox}{Number of ICs, and the number of data locations.}

\item{final}{If TRUE, return cov_S. Default: \code{FALSE}}
}
\value{
List of length three: \code{mu_S} (QxV), \code{E_SSt} (QxQ), and \code{cov_S} (QxQxV).
}
\description{
Update S for VB FC Template ICA
}
\keyword{internal}
