% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{update_S}
\alias{update_S}
\title{Update S for VB FC Template ICA}
\usage{
update_S(mu_tau2, mu_A, cov_A, D_inv, D_inv_S, BOLD, ntime, nICs, nvox)
}
\arguments{
\item{mu_tau2, mu_A, cov_A, }{Most recent estimates of posterior moments for
these variables.}

\item{D_inv, D_inv_S}{Some pre-computed quantities.
#' @param BOLD (\eqn{T \times V} matrix) preprocessed fMRI data.}

\item{ntime, nICs, nvox}{Number of timepoints in data, number of ICs, and
the number of data locations.}
}
\value{
List of length two: \code{mu_S} (QxV) and \code{cov_S} (QxQxV).
}
\description{
Update S for VB FC Template ICA
}
\keyword{internal}
