% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{update_tau2}
\alias{update_tau2}
\title{Update tau for VB FC Template ICA}
\usage{
update_tau2(BOLD, mu_A, mu_S, cov_A, cov_S, prior_params, ntime, nvox)
}
\arguments{
\item{BOLD}{(\eqn{T \times V} matrix) preprocessed fMRI data.}

\item{mu_A, mu_S, cov_A, cov_S, }{Most recent estimates of posterior
moments for these variables.}

\item{prior_params}{Alpha and beta parameters of IG prior on \eqn{\tau^2}
(error variance).}

\item{ntime, nvox}{Number of timepoints in data and the number of data
locations.}
}
\value{
List of length two: \code{mu_A} and \code{cov_A}.
}
\description{
Update tau for VB FC Template ICA
}
\keyword{internal}
