% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateTheta.R
\name{UpdateTheta}
\alias{UpdateTheta}
\alias{UpdateTheta.spatial}
\alias{UpdateTheta.independent}
\title{Parameter Estimates in EM Algorithm for Template ICA Model}
\usage{
UpdateTheta.spatial(template_mean, template_var, spde, BOLD, theta, C_diag,
  common_smoothness = TRUE, verbose = FALSE)

UpdateTheta.independent(template_mean, template_var, BOLD, theta, C_diag)
}
\arguments{
\item{template_mean}{mean maps for each IC in template}

\item{template_var}{between-subject variance maps for each IC in template}

\item{spde}{NULL for spatial independence model, otherwise SPDE object representing spatial prior on deviations.}

\item{BOLD}{dimension-reduced fMRI data}

\item{theta}{A list of current parameter estimates (mixing matrix A, noise variance nu0_sq and (for spatial model) SPDE parameters kappa)}

\item{C_diag}{(Qx1) diagonal elements of matrix proportional to residual variance.}

\item{common_smoothness}{If TRUE, use the common smoothness version of the spatial template ICA model, which assumes that all IC's have the same smoothness parameter, \eqn{\kappa}}

\item{verbose}{If TRUE, print progress updates for slow steps.}
}
\value{
An updated list of parameter estimates, theta
}
\description{
Parameter Estimates in EM Algorithm for Template ICA Model
}
