% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_decomp.R
\name{var_decomp}
\alias{var_decomp}
\title{Compute variance decomposition}
\usage{
var_decomp(x)
}
\arguments{
\item{x}{measurements by subjects by variables}
}
\value{
The variance decomposition
}
\description{
Compute variance decomposition
}
\keyword{internal}
