% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{VB_FCtemplateICA}
\alias{VB_FCtemplateICA}
\title{VB_FCtemplateICA}
\usage{
VB_FCtemplateICA(
  template_mean,
  template_var,
  template_FC,
  method_FC = c("VB1", "VB2"),
  prior_params = c(0.001, 0.001),
  BOLD,
  A0,
  S0,
  S0_var,
  maxiter = 100,
  miniter = 3,
  epsilon = 0.001,
  eps_inter = NULL,
  verbose = FALSE
)
}
\arguments{
\item{template_mean}{(\eqn{V \times Q} matrix) mean maps for each IC in the
template, where \eqn{Q} is the number of ICs, and \eqn{V=nvox} is the number
of data locations.}

\item{template_var}{(\eqn{V \times Q} matrix) between-subject variance maps
for each IC in the template.}

\item{template_FC}{(list) Parameters of functional connectivity template.}

\item{method_FC}{Variational Bayes (VB) method for FC template ICA model:
\code{"VB1"} (default) uses a conjugate Inverse-Wishart prior for the cor(A);
\code{"VB2"} draws samples from p(cor(A)) to emulate the population distribution
using a combination of Cholesky, SVD, and random pivoting.}

\item{prior_params}{Alpha and beta parameters of IG prior on \eqn{\tau^2}
(error variance). Default: \code{0.001} for both.}

\item{BOLD}{(\eqn{V \times T} matrix) preprocessed fMRI data.}

\item{A0, S0, S0_var}{Initial guesses at latent variables: \code{A} (\eqn{TxQ}
mixing matrix), \code{S} (\eqn{QxV} matrix of spatial ICs), and
variance matrix \code{S0_var}.}

\item{maxiter}{Maximum number of VB iterations. Default: \code{100}.}

\item{miniter}{Minimum number of VB iterations. Default: \code{3}.}

\item{epsilon}{Smallest proportion change in parameter estimates between iterations.
Default: \code{0.001}.}

\item{eps_inter}{Intermediate values of epsilon at which to save results (used
to assess benefit of more stringent convergence rules). Default:
\code{NULL} (do not save). These values should be in decreasing order
(larger to smaller error) and all values should be between zero and
\code{epsilon}.}

\item{verbose}{If \code{TRUE}, display progress of algorithm.
Default: \code{FALSE}.}
}
\value{
A list of computed values, including the final parameter estimates.
}
\description{
VB Algorithm for FC Template ICA Model
}
\keyword{internal}
