% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_FCtemplateICA.R
\name{VB_FCtemplateICA}
\alias{VB_FCtemplateICA}
\title{VB_FCtemplateICA}
\usage{
VB_FCtemplateICA(
  template_mean,
  template_var,
  template_FC,
  prior_params = c(0.001, 0.001),
  BOLD,
  A0,
  S0,
  S0_var,
  maxiter = 100,
  epsilon = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{template_mean}{(\eqn{V \times Q} matrix) mean maps for each IC in the
template, where \eqn{Q} is the number of ICs, and \eqn{V=nvox} is the number
of data locations.}

\item{template_var}{(\eqn{V \times Q} matrix) between-subject variance maps
for each IC in the template.}

\item{template_FC}{(list) Parameters of functional connectivity template.}

\item{prior_params}{Alpha and beta parameters of IG prior on \eqn{\tau^2}
(error variance). Default: \code{0.001} for both.}

\item{BOLD}{(\eqn{V \times T} matrix) preprocessed fMRI data.}

\item{A0, S0, S0_var}{Initial guesses at latent variables: \code{A} (\eqn{TxQ}
mixing matrix), \code{S} (\eqn{QxV} matrix of spatial ICs), and
covariance matrix \code{S0_var}.}

\item{maxiter}{Maximum number of VB iterations. Default: \code{100}.}

\item{epsilon}{Smallest proportion change in parameter estimates between
iterations. Default: \code{0.001}.}

\item{verbose}{If \code{TRUE}, display progress of algorithm. Default:
\code{FALSE}.}
}
\value{
Large list of results.
}
\description{
VB Algorithm for FC Template ICA Model
}
\keyword{internal}
